refactor(painters/dom): extract run rendering into runs/ module
Pull line- and run-level rendering out of renderer.ts (~2500 lines) into
a dedicated runs/ directory split by concern: text, image, math, tab,
field annotation, links, SDT, tracked changes, formatting marks, and the
top-level render-line/render-run orchestrators. Run hash helpers move
from paragraph-hash-utils.ts to runs/hash.ts. Public re-exports
(RenderedLineInfo, sanitizeUrl, linkMetrics, applyRunDataAttributes)
now resolve through runs/index.ts.

No behavior change — pure colocation of the run pipeline ahead of
unifying body and table-cell paragraph rendering.
