+++ packages/layout-engine/painters/dom/src/index.ts
@@ -40,10 +40,11 @@ export type {
   PaintSnapshotImageEntity,
   PaintSnapshotEntities,
 } from './renderer.js';
-export type { DomPainterInput, PositionMapping, RenderedLineInfo } from './renderer.js';
+export type { DomPainterInput, PositionMapping } from './renderer.js';
+export type { RenderedLineInfo } from './runs/index.js';
 
 // Re-export utility functions for testing
-export { sanitizeUrl, linkMetrics, applyRunDataAttributes } from './renderer.js';
+export { sanitizeUrl, linkMetrics, applyRunDataAttributes } from './runs/index.js';
 
 export { applySquareWrapExclusionsToLines } from './utils/anchor-helpers';
 export { buildImagePmSelector, buildInlineImagePmSelector } from './utils/image-selectors.js';
