+++ packages/layout-engine/painters/dom/src/renderer.ts
@@ -6,10 +6,8 @@ import type {
   DrawingFragment,
   DrawingGeometry,
   DropCapDescriptor,
-  FieldAnnotationRun,
   FlowBlock,
   FlowMode,
-  FlowRunLink,
   Fragment,
   GradientFill,
   ImageBlock,
@@ -37,10 +35,7 @@ import type {
   TableCellAttrs,
   TableFragment,
   TableMeasure,
-  MathRun,
   TextRun,
-  TrackedChangeKind,
-  TrackedChangesMode,
   VectorShapeDrawing,
   VectorShapeStyle,
   ResolvedLayout,
@@ -56,35 +51,18 @@ import type {
 import {
   LAYOUT_BOUNDARY_SCHEMA,
   buildLayoutSourceIdentityForFragment,
-  calculateJustifySpacing,
-  computeLinePmRange,
   expandRunsForInlineNewlines,
   getCellSpacingPx,
   getParagraphInlineDirection,
   normalizeColumnLayout,
-  normalizeBaselineShift,
-  resolveBaseFontSizeForVerticalText,
-  shouldApplyJustify,
-  sliceRunsForLine,
-  SPACE_CHARS,
 } from '@superdoc/contracts';
 import { DATASET_KEYS, decodeLayoutStoryDataset, encodeLayoutStoryDataset } from '@superdoc/dom-contract';
-import { toCssFontFamily } from '@superdoc/font-utils';
 import { getPresetShapeSvg } from '@superdoc/preset-geometry';
 import { encodeTooltip, sanitizeHref } from '@superdoc/url-validation';
 import { DOM_CLASS_NAMES } from './constants.js';
 import { createChartElement as renderChartToElement } from './chart-renderer.js';
-import {
-  getRunBooleanProp,
-  getRunNumberProp,
-  getRunStringProp,
-  getRunUnderlineColor,
-  getRunUnderlineStyle,
-  hashCellBorders,
-  hashParagraphBorders,
-  hashTableBorders,
-} from './paragraph-hash-utils.js';
-import { assertFragmentPmPositions, assertPmPositions } from './pm-position-validation.js';
+import { hashCellBorders, hashParagraphBorders, hashTableBorders } from './paragraph-hash-utils.js';
+import { assertFragmentPmPositions } from './pm-position-validation.js';
 import { createRulerElement, ensureRulerStyles, generateRulerDefinitionFromPx } from './ruler/index.js';
 import {
   BROWSER_DEFAULT_FONT_SIZE,
@@ -100,7 +78,6 @@ import {
   ensureSdtContainerStyles,
   ensureTrackChangeStyles,
   fragmentStyles,
-  lineStyles,
   pageStyles,
   spreadStyles,
   type PageStyles,
@@ -114,14 +91,18 @@ import {
 } from '@superdoc/common/list-marker-utils';
 import { applySdtContainerStyling, shouldRebuildForSdtBoundary, type SdtBoundaryOptions } from './utils/sdt-helpers.js';
 import { computeBetweenBorderFlags, type BetweenBorderInfo } from './paragraph/borders/index.js';
-import {
-  applyRtlStyles,
-  shouldUseSegmentPositioning,
-  resolveRunDirectionAttribute,
-  normalizeRtlDateTokenForWordParity,
-} from './features/inline-direction/index.js';
-import { convertOmmlToMathml } from './features/math/index.js';
 import { renderParagraphContent } from './paragraph/renderParagraphContent.js';
+import { renderLine as renderRunLine } from './runs/render-line.js';
+import type { RunRenderContext } from './runs/types.js';
+import {
+  getRunBooleanProp,
+  getRunNumberProp,
+  getRunStringProp,
+  getRunUnderlineColor,
+  getRunUnderlineStyle,
+} from './runs/hash.js';
+import { buildImageFilters } from './runs/image-run.js';
+import { applyTrackedChangeDecorations, resolveTrackedChangesConfig } from './runs/tracked-changes.js';
 
 type LineEnd = {
   type?: string;
@@ -827,449 +808,9 @@ function collectLineTabsForSnapshot(lineEl: HTMLElement): PaintSnapshotTabStyle[
 const DEFAULT_PAGE_HEIGHT_PX = 1056;
 /** Default gap used when virtualization is enabled (kept in sync with PresentationEditor layout defaults). */
 const DEFAULT_VIRTUALIZED_PAGE_GAP = 72;
-// Comment highlight color tokens moved to CommentHighlightDecorator (super-editor).
-
-type LinkRenderData = {
-  href?: string;
-  target?: string;
-  rel?: string;
-  tooltip?: string | null;
-  dataset?: Record<string, string>;
-  blocked: boolean;
-};
-
-const LINK_DATASET_KEYS = {
-  blocked: 'linkBlocked',
-  docLocation: 'linkDocLocation',
-  history: 'linkHistory',
-  rId: 'linkRid',
-  truncated: 'linkTooltipTruncated',
-} as const;
-
-const MAX_HREF_LENGTH = 2048;
-
-const SAFE_ANCHOR_PATTERN = /^[A-Za-z0-9._-]+$/;
-
-/**
- * Maximum allowed length for data URLs (10MB).
- * Prevents denial of service attacks from extremely large embedded images.
- */
-const MAX_DATA_URL_LENGTH = 10 * 1024 * 1024; // 10MB
-
-/**
- * Regular expression to validate data URL format for images.
- * Only allows common, safe image MIME types with base64 encoding.
- * Prevents XSS and malformed data URL attacks.
- */
-const VALID_IMAGE_DATA_URL = /^data:image\/(png|jpeg|jpg|gif|svg\+xml|webp|bmp|ico|tiff?);base64,/i;
 const SVG_NS = 'http://www.w3.org/2000/svg';
 const WORDART_LINE_FILL_RATIO = 0.9;
-
-/**
- * Maximum resize multiplier for image metadata.
- * Images can be resized up to 3x their original dimensions.
- */
-const MAX_RESIZE_MULTIPLIER = 3;
-
-/**
- * Fallback maximum dimension for image resizing when original size is small.
- * Ensures images can be resized to at least 1000px even if original is smaller.
- */
-const FALLBACK_MAX_DIMENSION = 1000;
-
-/**
- * Minimum image dimension in pixels.
- * Ensures images remain visible and interactive during resizing.
- */
-const MIN_IMAGE_DIMENSION = 20;
-
-/**
- * Pattern to detect ambiguous link text that doesn't convey destination (WCAG 2.4.4).
- * Matches common generic phrases like "click here", "read more", etc.
- */
-const AMBIGUOUS_LINK_PATTERNS = /^(click here|read more|more|link|here|this|download|view)$/i;
-
-/**
- * Hyperlink rendering metrics for observability.
- * Tracks sanitization, blocking, and security-related events.
- */
-const linkMetrics = {
-  sanitized: 0,
-  blocked: 0,
-  invalidProtocol: 0,
-  homographWarnings: 0,
-
-  reset() {
-    this.sanitized = 0;
-    this.blocked = 0;
-    this.invalidProtocol = 0;
-    this.homographWarnings = 0;
-  },
-
-  getMetrics() {
-    return {
-      'hyperlink.sanitized.count': this.sanitized,
-      'hyperlink.blocked.count': this.blocked,
-      'hyperlink.invalid_protocol.count': this.invalidProtocol,
-      'hyperlink.homograph_warnings.count': this.homographWarnings,
-    };
-  },
-};
-
-// Export for testing/monitoring
-export { linkMetrics };
-
-const TRACK_CHANGE_BASE_CLASS: Record<TrackedChangeKind, string> = {
-  insert: 'track-insert-dec',
-  delete: 'track-delete-dec',
-  format: 'track-format-dec',
-};
-// TRACK_CHANGE_FOCUSED_CLASS moved to CommentHighlightDecorator (super-editor).
-
-const TRACK_CHANGE_MODIFIER_CLASS: Record<TrackedChangeKind, Record<TrackedChangesMode, string | undefined>> = {
-  insert: {
-    review: 'highlighted',
-    original: 'hidden',
-    final: 'normal',
-    off: undefined,
-  },
-  delete: {
-    review: 'highlighted',
-    original: 'normal',
-    final: 'hidden',
-    off: undefined,
-  },
-  format: {
-    review: 'highlighted',
-    original: 'before',
-    final: 'normal',
-    off: undefined,
-  },
-};
-
-type TrackedChangesRenderConfig = {
-  mode: TrackedChangesMode;
-  enabled: boolean;
-};
-
-/**
- * Sanitize a URL to prevent XSS attacks.
- * Only allows http, https, mailto, tel, and internal anchors.
- *
- * @param href - The URL to sanitize
- * @returns Sanitized URL or null if blocked
- */
-export function sanitizeUrl(href: string): string | null {
-  if (typeof href !== 'string') return null;
-  const sanitized = sanitizeHref(href);
-  return sanitized?.href ?? null;
-}
-
-const LINK_TARGET_SET = new Set(['_blank', '_self', '_parent', '_top']);
-
-/**
- * Normalize and validate an anchor fragment identifier for use in hyperlinks.
- * Strips leading '#' if present and validates against safe character pattern.
- *
- * @param value - Raw anchor string (with or without leading '#')
- * @returns Normalized anchor with leading '#' (e.g., '#section-1'), or null if invalid
- *
- * @remarks
- * SECURITY: Only allows safe characters (A-Z, a-z, 0-9, ., _, -) to prevent HTML attribute injection.
- * Rejects characters like quotes, angle brackets, colons, and spaces that could break HTML structure
- * or enable XSS attacks when used in href attributes.
- *
- * @example
- * normalizeAnchor('section-1') // Returns: '#section-1'
- * normalizeAnchor('#bookmark') // Returns: '#bookmark'
- * normalizeAnchor('unsafe<script>') // Returns: null
- * normalizeAnchor('  whitespace  ') // Returns: null
- */
-const normalizeAnchor = (value: string | null | undefined): string | null => {
-  if (typeof value !== 'string') return null;
-  const trimmed = value.trim();
-  if (!trimmed) return null;
-
-  // Remove leading # if present, then validate
-  const anchor = trimmed.startsWith('#') ? trimmed.slice(1) : trimmed;
-
-  // SECURITY: Only allow safe characters to prevent attribute injection
-  // Rejects characters like quotes, angle brackets, spaces that could break HTML
-  if (!SAFE_ANCHOR_PATTERN.test(anchor)) {
-    return null;
-  }
-
-  return `#${anchor}`;
-};
-
-/**
- * Check if a fragment string contains only safe anchor characters.
- * Safe characters are alphanumeric, dots, underscores, and hyphens.
- *
- * @param {string} fragment - Fragment to validate
- * @returns {boolean} True if fragment matches safe pattern
- * @private
- */
-const isValidSafeFragment = (fragment: string): boolean => {
-  return SAFE_ANCHOR_PATTERN.test(fragment);
-};
-
-type ImageFilterSource = Pick<ImageBlock, 'grayscale' | 'gain' | 'blacklevel' | 'lum'>;
-
-const clampLumUnit = (value: number): number => {
-  return Math.max(-100000, Math.min(100000, value));
-};
-
-const parseVmlFixedFraction = (value: string | number | undefined): number | null => {
-  if (typeof value === 'number' && Number.isFinite(value)) {
-    return value;
-  }
-  if (typeof value !== 'string' || value.length === 0) {
-    return null;
-  }
-
-  if (value.endsWith('f')) {
-    const raw = Number.parseInt(value.slice(0, -1), 10);
-    return Number.isFinite(raw) ? raw / 65536 : null;
-  }
-
-  const parsed = Number.parseFloat(value);
-  return Number.isFinite(parsed) ? parsed : null;
-};
-
-const buildImageFilters = (source: ImageFilterSource): string[] => {
-  const filters: string[] = [];
-
-  if (source.grayscale) {
-    filters.push('grayscale(100%)');
-  }
-
-  if (source.gain != null || source.blacklevel != null) {
-    const gain = parseVmlFixedFraction(source.gain);
-    const blacklevel = parseVmlFixedFraction(source.blacklevel);
-
-    if (gain != null) {
-      const contrast = Math.max(0, gain);
-      if (contrast > 0) {
-        filters.push(`contrast(${contrast})`);
-      }
-    }
-
-    if (blacklevel != null) {
-      // CSS has no black-point control, so approximate VML blacklevel with a linear
-      // brightness shift using the same 0..32767 range Word's watermark UI uses.
-      const brightness = Math.max(0, 1 + blacklevel * (65536 / 32767));
-      if (brightness > 0) {
-        filters.push(`brightness(${brightness})`);
-      }
-    }
-  }
-
-  if (source.lum) {
-    // a:lum uses ST_FixedPercentage values expressed in thousandths of a percent.
-    // Convert those percentage deltas into CSS filter multipliers.
-    const contrastValue = typeof source.lum.contrast === 'number' ? clampLumUnit(source.lum.contrast) : null;
-    const brightValue = typeof source.lum.bright === 'number' ? clampLumUnit(source.lum.bright) : null;
-
-    if (contrastValue != null) {
-      const contrast = Math.max(0, 1 + contrastValue / 100000);
-      if (contrast >= 0) {
-        filters.push(`contrast(${contrast})`);
-      }
-    }
-
-    if (brightValue != null) {
-      const brightness = Math.max(0, 1 + brightValue / 100000);
-      if (brightness >= 0) {
-        filters.push(`brightness(${brightness})`);
-      }
-    }
-  }
-
-  return filters;
-};
-
-/**
- * URL-encode a fragment string for use in a URL hash.
- * Returns null if encoding fails (rare edge case).
- *
- * @param {string} fragment - Fragment to encode
- * @returns {string | null} Encoded fragment or null if encoding fails
- * @private
- */
-const encodeFragment = (fragment: string): string | null => {
-  try {
-    return encodeURIComponent(fragment);
-  } catch {
-    return null;
-  }
-};
-
-/**
- * Append a document location fragment to an href.
- * CRITICAL FIX: URL-encode unsafe characters instead of destroying the entire href.
- *
- * @param href - Base URL or null
- * @param docLocation - Fragment identifier to append
- * @returns Combined URL with fragment, or original href if fragment is invalid
- */
-const appendDocLocation = (href: string | null, docLocation?: string | null): string | null => {
-  if (!docLocation?.trim()) return href;
-
-  const fragment = docLocation.trim();
-  if (href?.includes('#')) return href;
-
-  const encoded = isValidSafeFragment(fragment) ? fragment : encodeFragment(fragment);
-
-  if (!encoded) return href;
-  return href ? `${href}#${encoded}` : `#${encoded}`;
-};
-
-/**
- * Build HTML data-* attributes object from hyperlink metadata for version 2 links.
- * Extracts relationship ID, document location fragment, and history preferences from link object.
- *
- * @param link - Flow run link object containing hyperlink metadata
- * @returns Record of data attribute keys and string values to be applied to anchor element
- *
- * @remarks
- * Only processes version 2 links (Office Open XML format). Version 1 links return empty object.
- * All dataset values are converted to strings for DOM compatibility.
- *
- * @example
- * buildLinkDataset({
- *   version: 2,
- *   rId: 'rId5',
- *   docLocation: 'bookmark1',
- *   history: true
- * })
- * // Returns: { rId: 'rId5', docLocation: 'bookmark1', history: 'true' }
- */
-const buildLinkDataset = (link: FlowRunLink): Record<string, string> => {
-  const dataset: Record<string, string> = {};
-  if (link.version === 2) {
-    if (link.rId) dataset[LINK_DATASET_KEYS.rId] = link.rId;
-    if (link.docLocation) dataset[LINK_DATASET_KEYS.docLocation] = link.docLocation;
-    if (typeof link.history === 'boolean') dataset[LINK_DATASET_KEYS.history] = String(link.history);
-  }
-  return dataset;
-};
-
-/**
- * Resolve the appropriate target attribute for a hyperlink anchor element.
- * Validates user-specified targets and auto-sets '_blank' for external HTTP(S) links.
- *
- * @param link - Flow run link object potentially containing target preference
- * @param sanitized - Sanitized URL metadata containing protocol information, or null if sanitization failed
- * @returns Valid target string ('_blank', '_self', '_parent', '_top') or undefined if not applicable
- *
- * @remarks
- * Target resolution follows this priority:
- * 1. If link.target is specified and valid (in LINK_TARGET_SET), use it
- * 2. If URL is external (http/https protocol), default to '_blank' for security
- * 3. Otherwise, return undefined (browser default behavior)
- *
- * @example
- * resolveLinkTarget(
- *   { target: '_self' },
- *   { protocol: 'https', href: 'https://example.com', isExternal: true }
- * ) // Returns: '_self' (user preference honored)
- *
- * resolveLinkTarget(
- *   {},
- *   { protocol: 'https', href: 'https://example.com', isExternal: true }
- * ) // Returns: '_blank' (external link default)
- */
-const resolveLinkTarget = (
-  link: FlowRunLink,
-  sanitized?: ReturnType<typeof sanitizeHref> | null,
-): string | undefined => {
-  if (link.target && LINK_TARGET_SET.has(link.target)) {
-    return link.target;
-  }
-  if (sanitized && (sanitized.protocol === 'http' || sanitized.protocol === 'https')) {
-    return '_blank';
-  }
-  return undefined;
-};
-
-/**
- * Resolve the rel attribute value for a hyperlink, combining user-specified relationships
- * with security-critical values for external links.
- *
- * @param link - Flow run link object potentially containing rel preference (space-separated string)
- * @param target - Resolved target attribute value (e.g., '_blank', '_self')
- * @returns Space-separated rel values, or undefined if no rel values apply
- *
- * @remarks
- * SECURITY: Automatically adds 'noopener noreferrer' for target='_blank' links to prevent:
- * - Tabnabbing attacks (window.opener access)
- * - Referrer leakage to external sites
- *
- * User-specified rel values are parsed from link.rel (whitespace-separated string),
- * deduplicated, and merged with security values.
- *
- * @example
- * resolveLinkRel(
- *   { rel: 'nofollow external' },
- *   '_blank'
- * ) // Returns: 'nofollow external noopener noreferrer'
- *
- * resolveLinkRel(
- *   { rel: 'nofollow  noopener  ' },
- *   '_blank'
- * ) // Returns: 'nofollow noopener noreferrer' (deduplicated)
- *
- * resolveLinkRel({}, '_self') // Returns: undefined
- */
-const resolveLinkRel = (link: FlowRunLink, target?: string): string | undefined => {
-  const relValues = new Set<string>();
-  if (typeof link.rel === 'string' && link.rel.trim()) {
-    link.rel
-      .trim()
-      .split(/\s+/)
-      .forEach((value) => {
-        if (value) relValues.add(value);
-      });
-  }
-  if (target === '_blank') {
-    relValues.add('noopener');
-    relValues.add('noreferrer');
-  }
-  if (relValues.size === 0) {
-    return undefined;
-  }
-  return Array.from(relValues).join(' ');
-};
-
-/**
- * Apply data-* attributes to an HTML element from a dataset object.
- * Safely assigns dataset properties while filtering out null/undefined values.
- *
- * @param element - Target HTML element to receive data attributes
- * @param dataset - Object mapping data attribute keys to string values
- *
- * @remarks
- * Uses the element.dataset API which automatically prefixes keys with 'data-'.
- * Only assigns non-null, non-undefined values to prevent empty attributes.
- *
- * @example
- * const anchor = document.createElement('a');
- * applyLinkDataset(anchor, {
- *   rId: 'rId5',
- *   docLocation: 'bookmark1',
- *   history: 'true'
- * });
- * // Resulting HTML: <a data-r-id="rId5" data-doc-location="bookmark1" data-history="true"></a>
- */
-const applyLinkDataset = (element: HTMLElement, dataset?: Record<string, string>): void => {
-  if (!dataset) return;
-  Object.entries(dataset).forEach(([key, value]) => {
-    if (value != null) {
-      element.dataset[key] = value;
-    }
-  });
-};
+// Comment highlight color tokens moved to CommentHighlightDecorator (super-editor).
 
 /**
  * DOM-based document painter that renders layout fragments to HTML elements.
@@ -4629,1062 +4170,6 @@ export class DomPainter {
     }
   }
 
-  /**
-   * Extract link data from a run, including sanitization.
-   * @returns Sanitized link data or null if invalid/missing
-   */
-  private extractLinkData(run: Run): LinkRenderData | null {
-    if (run.kind === 'tab' || run.kind === 'image' || run.kind === 'lineBreak' || run.kind === 'math') {
-      return null;
-    }
-    const link = (run as TextRun).link as FlowRunLink | undefined;
-    if (!link) {
-      return null;
-    }
-    return this.buildLinkRenderData(link);
-  }
-
-  private buildLinkRenderData(link: FlowRunLink): LinkRenderData | null {
-    const dataset = buildLinkDataset(link);
-    const sanitized = typeof link.href === 'string' ? sanitizeHref(link.href) : null;
-    const anchorHref = normalizeAnchor(link.anchor ?? link.name ?? '');
-    let href: string | null = sanitized?.href ?? anchorHref;
-    if (link.version === 2) {
-      href = appendDocLocation(href, link.docLocation ?? null);
-    }
-
-    // Track metrics: successful sanitization
-    if (sanitized) {
-      linkMetrics.sanitized++;
-
-      // Check for homograph if hostname has non-ASCII (in raw href before URL parsing)
-      if (sanitized.href && typeof link.href === 'string') {
-        const hostStartIndex = link.href.indexOf('://') + 3;
-        let hostEndIndex = link.href.indexOf('/', hostStartIndex);
-        if (hostEndIndex === -1) {
-          hostEndIndex = link.href.indexOf('?', hostStartIndex);
-        }
-        if (hostEndIndex === -1) {
-          hostEndIndex = link.href.indexOf('#', hostStartIndex);
-        }
-        if (hostEndIndex === -1) {
-          hostEndIndex = link.href.length;
-        }
-        const rawHostname = link.href.slice(hostStartIndex, hostEndIndex);
-        if (rawHostname && /[^\x00-\x7F]/.test(rawHostname)) {
-          linkMetrics.homographWarnings++;
-        }
-      }
-    }
-
-    // Defense-in-depth: Enforce maximum URL length even if sanitization was bypassed
-    if (sanitized && sanitized.href.length > MAX_HREF_LENGTH) {
-      console.warn(`[DomPainter] Rejecting URL exceeding ${MAX_HREF_LENGTH} characters`);
-      linkMetrics.blocked++;
-      return { blocked: true, dataset: { [LINK_DATASET_KEYS.blocked]: 'true' } };
-    }
-
-    if (!href) {
-      if (typeof link.href === 'string' && link.href.trim()) {
-        dataset[LINK_DATASET_KEYS.blocked] = 'true';
-        console.warn(`[DomPainter] Blocked potentially unsafe URL: ${link.href.slice(0, 50)}`);
-        linkMetrics.blocked++;
-        // Track invalid protocol if sanitized was null
-        if (!sanitized) {
-          linkMetrics.invalidProtocol++;
-        }
-        return { blocked: true, dataset };
-      }
-      // Check if there was an anchor/name that failed validation
-      const hadAnchor = (link.anchor ?? link.name ?? null) != null;
-      if (Object.keys(dataset).length > 0 || hadAnchor) {
-        dataset[LINK_DATASET_KEYS.blocked] = 'true';
-        linkMetrics.blocked++;
-        return { blocked: true, dataset };
-      }
-      return null;
-    }
-
-    const target = resolveLinkTarget(link, sanitized);
-    const rel = resolveLinkRel(link, target);
-    const tooltipSource = link.version === 2 ? (link.tooltip ?? link.title) : link.title;
-    const tooltipResult = tooltipSource ? encodeTooltip(tooltipSource) : null;
-    // Use raw text - browser will escape when setting attribute
-    const tooltip = tooltipResult?.text ?? null;
-
-    // Signal when tooltip is truncated
-    if (tooltipResult?.wasTruncated) {
-      dataset[LINK_DATASET_KEYS.truncated] = 'true';
-    }
-
-    return {
-      href,
-      target,
-      rel,
-      tooltip,
-      dataset: Object.keys(dataset).length > 0 ? dataset : undefined,
-      blocked: false,
-    };
-  }
-
-  /**
-   * Apply tooltip accessibility using aria-describedby for better screen reader support.
-   * Creates a visually-hidden element containing the tooltip text and links it to the anchor.
-   *
-   * @param elem - The anchor element to enhance
-   * @param tooltip - The tooltip text to make accessible
-   * @returns The unique ID generated for this link
-   */
-  private applyTooltipAccessibility(elem: HTMLAnchorElement, tooltip: string | null): string {
-    const linkId = `superdoc-link-${++this.linkIdCounter}`;
-    elem.id = linkId;
-
-    if (!tooltip || !this.doc) return linkId;
-
-    // Keep title attribute for visual tooltip (browser default)
-    elem.setAttribute('title', tooltip);
-
-    // Create visually-hidden element for screen readers
-    const descId = `link-desc-${linkId}`;
-    const descElem = this.doc.createElement('span');
-    descElem.id = descId;
-    descElem.className = 'superdoc-sr-only'; // Screen reader only class
-    descElem.textContent = tooltip;
-
-    // Insert description element after the link
-    // Note: We'll insert it as a sibling in the parent line element
-    if (elem.parentElement) {
-      elem.parentElement.appendChild(descElem);
-      // Reference from link only if we successfully added the description element
-      elem.setAttribute('aria-describedby', descId);
-    } else {
-      // Element not yet in DOM - accessibility feature will degrade gracefully
-      // The title attribute will still provide tooltip functionality
-      console.warn('[DomPainter] Unable to add aria-describedby for tooltip (element not in DOM)');
-    }
-
-    return linkId;
-  }
-
-  /**
-   * Enhance accessibility of a link element with ARIA labels and attributes.
-   * Adds descriptive ARIA labels for ambiguous text and target=_blank links (WCAG 2.4.4).
-   *
-   * @param elem - The anchor element to enhance
-   * @param linkData - Link metadata including href and target
-   * @param textContent - The visible link text to analyze for ambiguity
-   */
-  private enhanceAccessibility(elem: HTMLAnchorElement, linkData: LinkRenderData, textContent: string): void {
-    if (!linkData.href) return;
-
-    const trimmedText = textContent.trim().toLowerCase();
-
-    // Check if link text is ambiguous (e.g., "click here", "read more")
-    if (AMBIGUOUS_LINK_PATTERNS.test(trimmedText)) {
-      try {
-        const url = new URL(linkData.href);
-        const hostname = url.hostname.replace(/^www\./, '');
-
-        // Generate descriptive aria-label for screen readers
-        const ariaLabel = `${textContent.trim()} - ${hostname}`;
-        elem.setAttribute('aria-label', ariaLabel);
-        return; // Exit early since we've set the label
-      } catch {
-        // If URL parsing fails, add generic label
-        elem.setAttribute('aria-label', `${textContent.trim()} - external link`);
-        return;
-      }
-    }
-
-    // Add aria-label for external links without one (indicates new tab)
-    if (linkData.target === '_blank' && !elem.getAttribute('aria-label')) {
-      elem.setAttribute('aria-label', `${textContent.trim()} (opens in new tab)`);
-    }
-  }
-
-  /**
-   * Apply link attributes to an anchor element.
-   */
-  private applyLinkAttributes(elem: HTMLAnchorElement, linkData: LinkRenderData): void {
-    if (!linkData.href) return;
-    elem.href = linkData.href;
-    elem.classList.add('superdoc-link');
-
-    if (linkData.target) {
-      elem.target = linkData.target;
-    } else {
-      elem.removeAttribute('target');
-    }
-    if (linkData.rel) {
-      elem.rel = linkData.rel;
-    } else {
-      elem.removeAttribute('rel');
-    }
-    if (linkData.tooltip) {
-      elem.title = linkData.tooltip;
-    } else {
-      elem.removeAttribute('title');
-    }
-
-    // Explicitly set role for clarity (though <a> with href has implicit role="link")
-    elem.setAttribute('role', 'link');
-
-    // Ensure link is keyboard accessible (should be default for <a>, but verify)
-    elem.setAttribute('tabindex', '0');
-
-    // Note: Click handling is done via event delegation in EditorInputManager,
-    // not per-element handlers. This avoids duplicate event dispatching.
-  }
-
-  /**
-   * Render a single run as an HTML element (span or anchor).
-   */
-  /**
-   * Type guard to check if a run is an image run.
-   */
-  private isImageRun(run: Run): run is ImageRun {
-    return run.kind === 'image';
-  }
-
-  /**
-   * Type guard to check if a run is a line break run.
-   */
-  private isLineBreakRun(run: Run): run is import('@superdoc/contracts').LineBreakRun {
-    return run.kind === 'lineBreak';
-  }
-
-  /**
-   * Type guard to check if a run is a break run.
-   */
-  private isBreakRun(run: Run): run is import('@superdoc/contracts').BreakRun {
-    return run.kind === 'break';
-  }
-
-  /**
-   * Type guard to check if a run is a field annotation run.
-   */
-  private isFieldAnnotationRun(run: Run): run is FieldAnnotationRun {
-    return run.kind === 'fieldAnnotation';
-  }
-
-  /**
-   * Type guard to check if a run is a math run.
-   */
-  private isMathRun(run: Run): run is MathRun {
-    return run.kind === 'math';
-  }
-
-  /**
-   * Render a math run as a MathML element wrapped in a span.
-   * Follows the same pattern as renderImageRun — sets explicit dimensions.
-   */
-  private renderMathRun(run: MathRun): HTMLElement | null {
-    if (!this.doc) return null;
-    const wrapper = this.doc.createElement('span');
-    wrapper.className = 'sd-math';
-    wrapper.style.display = 'inline-block';
-    wrapper.style.verticalAlign = 'middle';
-    // Let browser auto-size to MathML content; estimated dimensions are for layout only
-    wrapper.style.minWidth = `${run.width}px`;
-    wrapper.style.minHeight = `${run.height}px`;
-    // Restore font-size so the plain-text fallback renders at a reasonable size
-    // (the line container sets fontSize: 0 to eliminate the CSS strut). MathML
-    // has its own internal scaling, so this only matters for the textContent
-    // fallback path. run.height would make tall expressions (fractions, equation
-    // arrays) render at 80–100px — use the browser default instead.
-    wrapper.style.fontSize = BROWSER_DEFAULT_FONT_SIZE;
-    wrapper.dataset.layoutEpoch = String(this.layoutEpoch ?? 0);
-
-    const mathEl = convertOmmlToMathml(run.ommlJson, this.doc);
-    if (mathEl) {
-      wrapper.appendChild(mathEl);
-    } else {
-      // Fallback: render plain text content
-      wrapper.textContent = run.textContent || '';
-    }
-
-    if (run.pmStart != null) wrapper.dataset.pmStart = String(run.pmStart);
-    if (run.pmEnd != null) wrapper.dataset.pmEnd = String(run.pmEnd);
-
-    return wrapper;
-  }
-
-  private setTextContentWithFormattingSpaceMarks(element: HTMLElement, text: string): void {
-    if (!this.showFormattingMarks || !text.includes(' ') || !this.doc) {
-      element.textContent = text;
-      return;
-    }
-
-    element.textContent = '';
-    let chunkStart = 0;
-    for (let index = 0; index < text.length; index += 1) {
-      if (text[index] !== ' ') continue;
-
-      if (index > chunkStart) {
-        element.appendChild(this.doc.createTextNode(text.slice(chunkStart, index)));
-      }
-
-      const space = this.doc.createElement('span');
-      space.classList.add('superdoc-formatting-space-mark');
-      space.textContent = ' ';
-      element.appendChild(space);
-      chunkStart = index + 1;
-    }
-
-    if (chunkStart < text.length) {
-      element.appendChild(this.doc.createTextNode(text.slice(chunkStart)));
-    }
-  }
-
-  private findLastTextRun(runs: Run[]): { run: TextRun; index: number } | null {
-    for (let index = runs.length - 1; index >= 0; index -= 1) {
-      const run = runs[index];
-      if (run && (run.kind === 'text' || run.kind === undefined) && 'text' in run) {
-        return { run: run as TextRun, index };
-      }
-    }
-    return null;
-  }
-
-  private appendFormattingParagraphMark(
-    lineEl: HTMLElement,
-    line: Line,
-    runs: Run[],
-    leftOffsetPx: number,
-    availableWidth: number,
-    hasExplicitPositioning: boolean,
-  ): void {
-    if (!this.showFormattingMarks || !this.doc) return;
-    const lastRun = runs.length > 0 ? runs[runs.length - 1] : null;
-    if (lastRun) {
-      const lastRunIndex = runs.length - 1;
-      if (line.toRun < lastRunIndex) return;
-      if (
-        line.toRun === lastRunIndex &&
-        (lastRun.kind === 'text' || lastRun.kind === undefined) &&
-        'text' in lastRun &&
-        line.toChar < lastRun.text.length
-      ) {
-        return;
-      }
-    }
-
-    const lastTextRun = this.findLastTextRun(runs);
-
-    const mark = this.doc.createElement('span');
-    mark.classList.add('superdoc-formatting-paragraph-mark');
-    mark.setAttribute('aria-hidden', 'true');
-    mark.textContent = '¶';
-
-    const run = lastTextRun?.run;
-    if (run) {
-      if (run.fontFamily) {
-        mark.style.fontFamily = toCssFontFamily(run.fontFamily) ?? run.fontFamily;
-      }
-      if (typeof run.fontSize === 'number') {
-        mark.style.fontSize = `${run.fontSize}px`;
-      }
-      if (run.bold) {
-        mark.style.fontWeight = 'bold';
-      }
-      if (run.italic) {
-        mark.style.fontStyle = 'italic';
-      }
-      if (run.letterSpacing != null) {
-        mark.style.letterSpacing = `${run.letterSpacing}px`;
-      }
-    }
-    mark.style.lineHeight = `${line.lineHeight}px`;
-
-    const lineWidth = line.naturalWidth ?? line.width ?? 0;
-    const alignmentSlack = Math.max(0, availableWidth - lineWidth);
-    const textAlign = lineEl.style.textAlign;
-    const alignmentOffset =
-      !hasExplicitPositioning && textAlign === 'center'
-        ? alignmentSlack / 2
-        : !hasExplicitPositioning && textAlign === 'right'
-          ? alignmentSlack
-          : 0;
-    const isRtl = lineEl.dir === 'rtl' || lineEl.style.direction === 'rtl';
-    const visualTextEndOffset = isRtl ? alignmentOffset : alignmentOffset + lineWidth;
-    mark.style.left = `${Math.max(0, leftOffsetPx + visualTextEndOffset)}px`;
-    lineEl.appendChild(mark);
-  }
-
-  private renderRun(
-    run: Run,
-    context: FragmentRenderContext,
-    trackedConfig?: TrackedChangesRenderConfig,
-  ): HTMLElement | null {
-    // Handle ImageRun
-    if (this.isImageRun(run)) {
-      return this.renderImageRun(run);
-    }
-
-    // Handle FieldAnnotationRun - inline pill-styled form fields
-    if (this.isFieldAnnotationRun(run)) {
-      return this.renderFieldAnnotationRun(run);
-    }
-
-    // Handle MathRun - inline math rendered as MathML
-    if (this.isMathRun(run)) {
-      return this.renderMathRun(run);
-    }
-
-    // Handle LineBreakRun - line breaks are handled by the measurer creating new lines,
-    // so we don't render anything for them in the DOM. They exist in the run array for
-    // proper PM position tracking but don't need visual representation.
-    if (this.isLineBreakRun(run)) {
-      return null;
-    }
-
-    // Handle BreakRun - similar to LineBreakRun, breaks are handled by the measurer
-    if (this.isBreakRun(run)) {
-      return null;
-    }
-
-    // Handle TextRun
-    if (!('text' in run) || !run.text || !this.doc) {
-      return null;
-    }
-
-    const linkData = this.extractLinkData(run);
-    const isActiveLink = !!(linkData && !linkData.blocked && linkData.href);
-    const elem = isActiveLink ? this.doc.createElement('a') : this.doc.createElement('span');
-    const text = resolveRunText(run, context);
-    const textRun = run as TextRun;
-    const effectiveText =
-      textRun.bidi?.rtl === true && typeof text === 'string' ? normalizeRtlDateTokenForWordParity(text) : text;
-    this.setTextContentWithFormattingSpaceMarks(elem, effectiveText);
-
-    if (linkData?.dataset) {
-      applyLinkDataset(elem, linkData.dataset);
-    }
-    if (linkData?.blocked) {
-      elem.dataset[LINK_DATASET_KEYS.blocked] = 'true';
-      // For blocked links rendered as spans, set appropriate role
-      elem.setAttribute('role', 'text');
-      elem.setAttribute('aria-label', 'Invalid link - not clickable');
-    }
-    if (isActiveLink && linkData) {
-      this.applyLinkAttributes(elem as HTMLAnchorElement, linkData);
-      // Enhance accessibility with ARIA labels for ambiguous text
-      this.enhanceAccessibility(elem as HTMLAnchorElement, linkData, text);
-
-      // Note: Tooltip accessibility (aria-describedby) will be applied after
-      // the element is added to the DOM in renderLine, since it creates a sibling element
-      // Store tooltip for later processing
-      if (linkData.tooltip) {
-        this.pendingTooltips.set(elem, linkData.tooltip);
-      }
-    }
-
-    // Pass isLink flag to skip applying inline color/decoration styles for links
-    applyRunStyles(elem as HTMLElement, run, isActiveLink);
-    // SD-3098 Word-parity: run-level dir attribute decision lives in the
-    // inline-direction feature. See resolveRunDirectionAttribute for the
-    // decision table (rtl-tagged + content classification + date-like).
-    const dirAttr = resolveRunDirectionAttribute({
-      runText: textRun.text,
-      effectiveText,
-      isRtlTagged: textRun.bidi?.rtl === true,
-    });
-    if (dirAttr) {
-      elem.setAttribute('dir', dirAttr);
-    }
-    const commentAnnotations = textRun.comments;
-    const hasAnyComment = !!commentAnnotations?.length;
-    // Comment highlight styles are applied post-paint by CommentHighlightDecorator (super-editor).
-    // The painter only stamps metadata attributes below.
-    // We still need to preserve the comment ids
-    if (hasAnyComment) {
-      elem.dataset.commentIds = commentAnnotations.map((c) => c.commentId).join(',');
-      if (commentAnnotations.some((c) => c.internal)) {
-        elem.dataset.commentInternal = 'true';
-      }
-      // Per-comment internal flag so the editor-side decorator can pick the right color
-      const internalIds = commentAnnotations.filter((c) => c.internal).map((c) => c.commentId);
-      if (internalIds.length > 0) {
-        elem.dataset.commentInternalIds = internalIds.join(',');
-      }
-      // importedId aliases so the decorator can match by either ID
-      const importedEntries = commentAnnotations
-        .filter((c) => c.importedId && c.importedId !== c.commentId)
-        .map((c) => `${c.importedId}=${c.commentId}`);
-      if (importedEntries.length > 0) {
-        elem.dataset.commentImportedIds = importedEntries.join(',');
-      }
-      elem.classList.add('superdoc-comment-highlight');
-    }
-    // Ensure text renders above tab leaders (leaders are z-index: 0)
-    elem.style.zIndex = '1';
-    applyRunDataAttributes(elem as HTMLElement, (run as TextRun).dataAttrs);
-
-    // SD-2454: bookmark marker runs carry a data-bookmark-name attribute.
-    // Surface the bookmark name as a native `title` tooltip so hovering the
-    // opening bracket identifies which bookmark is being marked.
-    const bookmarkName = (run as TextRun).dataAttrs?.['data-bookmark-name'];
-    if (bookmarkName) {
-      (elem as HTMLElement).title = bookmarkName;
-    }
-
-    // Assert PM positions are present for cursor fallback
-    assertPmPositions(run, 'paragraph text run');
-
-    if (run.pmStart != null) elem.dataset.pmStart = String(run.pmStart);
-    if (run.pmEnd != null) elem.dataset.pmEnd = String(run.pmEnd);
-    elem.dataset.layoutEpoch = String(this.layoutEpoch);
-    if (trackedConfig) {
-      this.applyTrackedChangeDecorations(elem, run, trackedConfig);
-    }
-    this.applySdtDataset(elem, (run as TextRun).sdt);
-
-    return elem;
-  }
-
-  /**
-   * Renders an ImageRun as an inline <img> element.
-   *
-   * SECURITY NOTES:
-   * - Data URLs are validated against VALID_IMAGE_DATA_URL regex to ensure proper format
-   * - Size limit (MAX_DATA_URL_LENGTH) prevents DoS attacks from extremely large images
-   * - Only allows safe image MIME types (png, jpeg, gif, etc.) with base64 encoding
-   * - Non-data URLs are sanitized through sanitizeUrl to prevent XSS
-   *
-   * METADATA ATTRIBUTE:
-   * - Adds `data-image-metadata` attribute to enable interactive resizing via ImageResizeOverlay
-   * - Metadata includes: originalWidth, originalHeight, aspectRatio, min/max dimensions
-   * - Only added when run.width > 0 && run.height > 0 to prevent invalid metadata
-   * - Max dimensions: 3x original size or 1000px (whichever is larger)
-   * - Min dimensions: 20px to ensure visibility and interactivity
-   *
-   * @param run - The ImageRun to render containing image source, dimensions, and spacing
-   * @returns HTMLElement (img) or null if src is missing or invalid
-   *
-   * @example
-   * ```typescript
-   * // Valid data URL with metadata
-   * renderImageRun({ kind: 'image', src: 'data:image/png;base64,iVBORw...', width: 100, height: 100 })
-   * // Returns: <img> element with data-image-metadata attribute
-   *
-   * // Invalid dimensions - no metadata
-   * renderImageRun({ kind: 'image', src: 'data:image/png;base64,iVBORw...', width: 0, height: 0 })
-   * // Returns: <img> element WITHOUT data-image-metadata attribute
-   *
-   * // Invalid MIME type
-   * renderImageRun({ kind: 'image', src: 'data:text/html;base64,PHNjcmlwdD4...', width: 100, height: 100 })
-   * // Returns: null (blocked)
-   *
-   * // HTTP URL
-   * renderImageRun({ kind: 'image', src: 'https://example.com/image.png', width: 100, height: 100 })
-   * // Returns: <img> element (after sanitization) with data-image-metadata attribute
-   * ```
-   */
-  private renderImageRun(run: ImageRun): HTMLElement | null {
-    if (!this.doc || !run.src) {
-      return null;
-    }
-
-    const hasClipPath = typeof run.clipPath === 'string' && run.clipPath.trim().length > 0;
-
-    // Create img element
-    const img = this.doc.createElement('img');
-    img.classList.add(DOM_CLASS_NAMES.INLINE_IMAGE);
-
-    // Set source - validate data URLs with strict format and size checks
-    // Note: data: URLs are blocked by sanitizeUrl for hyperlinks (XSS risk),
-    // but are safe for <img> elements when properly validated
-    const isDataUrl = typeof run.src === 'string' && run.src.startsWith('data:');
-    if (isDataUrl) {
-      // SECURITY: Validate data URL format and size
-      if (run.src.length > MAX_DATA_URL_LENGTH) {
-        // Reject data URLs that are too large (DoS prevention)
-        return null;
-      }
-      if (!VALID_IMAGE_DATA_URL.test(run.src)) {
-        // Reject data URLs with invalid MIME types or encoding
-        return null;
-      }
-      img.src = run.src;
-    } else {
-      const sanitized = sanitizeUrl(run.src);
-      if (sanitized) {
-        img.src = sanitized;
-      } else {
-        // Invalid URL - return null
-        return null;
-      }
-    }
-
-    // Set dimensions: when we have clipPath we put img in a wrapper that has the layout size and overflow:hidden; img fills wrapper so cropped portion stays within after resize
-    if (!hasClipPath) {
-      img.width = run.width;
-      img.height = run.height;
-    } else {
-      Object.assign(img.style, {
-        width: '100%',
-        height: '100%',
-        maxWidth: '100%',
-        maxHeight: '100%',
-        boxSizing: 'border-box',
-        minWidth: '0',
-        minHeight: '0',
-      });
-    }
-    applyImageClipPath(img, run.clipPath);
-
-    // Add metadata for interactive image resizing (inline images)
-    // Only add metadata if dimensions are valid (positive, non-zero values)
-    if (run.width > 0 && run.height > 0) {
-      // This enables the ImageResizeOverlay to work with inline images
-      const aspectRatio = run.width / run.height;
-      const inlineImageMetadata = {
-        originalWidth: run.width,
-        originalHeight: run.height,
-        // Max dimensions: MAX_RESIZE_MULTIPLIER x original size or FALLBACK_MAX_DIMENSION, whichever is larger
-        // This provides generous constraints while preventing excessive scaling
-        maxWidth: Math.max(run.width * MAX_RESIZE_MULTIPLIER, FALLBACK_MAX_DIMENSION),
-        maxHeight: Math.max(run.height * MAX_RESIZE_MULTIPLIER, FALLBACK_MAX_DIMENSION),
-        aspectRatio,
-        // Min dimensions: MIN_IMAGE_DIMENSION to ensure images remain visible and interactive
-        minWidth: MIN_IMAGE_DIMENSION,
-        minHeight: MIN_IMAGE_DIMENSION,
-      };
-      img.setAttribute('data-image-metadata', JSON.stringify(inlineImageMetadata));
-    }
-
-    // Set alt text (required for accessibility)
-    img.alt = run.alt ?? '';
-
-    // Set title if present
-    if (run.title) {
-      img.title = run.title;
-    }
-
-    // Apply inline-block display
-    img.style.display = 'inline-block';
-
-    // When we use a wrapper (clipPath + positive dimensions), margins/verticalAlign/position/zIndex go on the wrapper only.
-    // When we don't use a wrapper (no clipPath, or clipPath with width/height 0), apply them on the img so layout is correct.
-    const useWrapper = hasClipPath && run.width > 0 && run.height > 0;
-    if (!useWrapper) {
-      // Apply vertical alignment (bottom-aligned to text baseline)
-      img.style.verticalAlign = run.verticalAlign ?? 'bottom';
-
-      // Apply spacing as CSS margins
-      if (run.distTop) {
-        img.style.marginTop = `${run.distTop}px`;
-      }
-      if (run.distBottom) {
-        img.style.marginBottom = `${run.distBottom}px`;
-      }
-      if (run.distLeft) {
-        img.style.marginLeft = `${run.distLeft}px`;
-      }
-      if (run.distRight) {
-        img.style.marginRight = `${run.distRight}px`;
-      }
-
-      // Position and z-index on the image only (not the line) so resize overlay can stack above.
-      img.style.position = 'relative';
-      img.style.zIndex = '1';
-      img.style.maxWidth = '100%';
-    }
-
-    // Apply rotation and flip transforms from OOXML a:xfrm
-    const transforms: string[] = [];
-
-    // Calculate translation offset to keep top-left corner fixed when rotating
-    if (run.rotation != null && run.rotation !== 0) {
-      const angleRad = (run.rotation * Math.PI) / 180;
-      const w = run.width;
-      const h = run.height;
-
-      // Calculate how much the top-left corner moves when rotating around center
-      // Top-left corner starts at (0, 0) in element space
-      // Center is at (w/2, h/2)
-      // After rotation, we need to translate to keep top-left at (0, 0)
-      const cosA = Math.cos(angleRad);
-      const sinA = Math.sin(angleRad);
-
-      // Position of top-left corner after rotation (relative to original top-left)
-      const newTopLeftX = (w / 2) * (1 - cosA) + (h / 2) * sinA;
-      const newTopLeftY = (w / 2) * sinA + (h / 2) * (1 - cosA);
-
-      transforms.push(`translate(${-newTopLeftX}px, ${-newTopLeftY}px)`);
-      transforms.push(`rotate(${run.rotation}deg)`);
-    }
-    if (run.flipH) {
-      transforms.push('scaleX(-1)');
-    }
-    if (run.flipV) {
-      transforms.push('scaleY(-1)');
-    }
-    if (transforms.length > 0) {
-      img.style.transform = transforms.join(' ');
-      img.style.transformOrigin = 'center';
-    }
-
-    const filters = buildImageFilters(run);
-    if (filters.length > 0) {
-      img.style.filter = filters.join(' ');
-    }
-
-    // Assert PM positions are present for cursor fallback
-    assertPmPositions(run, 'inline image run');
-
-    // When clipPath is set, scale makes the img paint outside its box;
-    // wrap in a clip container so only the cropped portion occupies space in the document.
-    // Wrapper size is the only layout box (position calculation uses run.width/run.height).
-    // PM position attributes go on the wrapper only so selection highlight and selection rects use the wrapper, not the scaled img.
-    // Skip wrapper when width or height is 0 (no layout box); img already has margins/verticalAlign/position/zIndex from above.
-    if (useWrapper) {
-      const wrapper = this.doc.createElement('span');
-      wrapper.classList.add(DOM_CLASS_NAMES.INLINE_IMAGE_CLIP_WRAPPER);
-      wrapper.style.display = 'inline-block';
-      wrapper.style.width = `${run.width}px`;
-      wrapper.style.height = `${run.height}px`;
-      wrapper.style.boxSizing = 'border-box';
-      wrapper.style.overflow = 'hidden';
-      wrapper.style.verticalAlign = run.verticalAlign ?? 'bottom';
-      if (run.distTop) wrapper.style.marginTop = `${run.distTop}px`;
-      if (run.distBottom) wrapper.style.marginBottom = `${run.distBottom}px`;
-      if (run.distLeft) wrapper.style.marginLeft = `${run.distLeft}px`;
-      if (run.distRight) wrapper.style.marginRight = `${run.distRight}px`;
-      wrapper.style.position = 'relative';
-      wrapper.style.zIndex = '1';
-      if (run.pmStart != null) wrapper.dataset.pmStart = String(run.pmStart);
-      if (run.pmEnd != null) wrapper.dataset.pmEnd = String(run.pmEnd);
-      wrapper.dataset.layoutEpoch = String(this.layoutEpoch);
-      this.applySdtDataset(wrapper, run.sdt);
-      if (run.dataAttrs) applyRunDataAttributes(wrapper, run.dataAttrs);
-      wrapper.appendChild(img);
-      return this.buildImageHyperlinkAnchor(wrapper, run.hyperlink, 'inline-block');
-    }
-
-    // Apply PM position tracking for cursor placement (only on img when not wrapped)
-    if (run.pmStart != null) {
-      img.dataset.pmStart = String(run.pmStart);
-    }
-    if (run.pmEnd != null) {
-      img.dataset.pmEnd = String(run.pmEnd);
-    }
-    img.dataset.layoutEpoch = String(this.layoutEpoch);
-
-    // Apply SDT metadata
-    this.applySdtDataset(img, run.sdt);
-
-    // Apply data attributes
-    if (run.dataAttrs) {
-      applyRunDataAttributes(img, run.dataAttrs);
-    }
-
-    const runClipPath = readClipPathValue((run as { clipPath?: unknown }).clipPath);
-    if (runClipPath && this.doc) {
-      img.style.clipPath = runClipPath;
-      img.style.display = 'block';
-      img.style.marginTop = '';
-      img.style.marginBottom = '';
-      img.style.marginLeft = '';
-      img.style.marginRight = '';
-      img.style.verticalAlign = '';
-      img.style.position = 'static';
-      img.style.zIndex = '';
-
-      const wrapper = this.doc.createElement('span');
-      wrapper.classList.add('superdoc-inline-image-clip-wrapper');
-      wrapper.style.display = 'inline-block';
-      wrapper.style.width = `${run.width}px`;
-      wrapper.style.height = `${run.height}px`;
-      wrapper.style.verticalAlign = run.verticalAlign ?? 'bottom';
-      wrapper.style.position = 'relative';
-      wrapper.style.zIndex = '1';
-      if (run.distTop) wrapper.style.marginTop = `${run.distTop}px`;
-      if (run.distBottom) wrapper.style.marginBottom = `${run.distBottom}px`;
-      if (run.distLeft) wrapper.style.marginLeft = `${run.distLeft}px`;
-      if (run.distRight) wrapper.style.marginRight = `${run.distRight}px`;
-
-      if (run.pmStart != null) {
-        wrapper.dataset.pmStart = String(run.pmStart);
-      }
-      if (run.pmEnd != null) {
-        wrapper.dataset.pmEnd = String(run.pmEnd);
-      }
-      wrapper.dataset.layoutEpoch = String(this.layoutEpoch);
-      this.applySdtDataset(wrapper, run.sdt);
-
-      wrapper.appendChild(img);
-      return this.buildImageHyperlinkAnchor(wrapper, run.hyperlink, 'inline-block');
-    }
-
-    return this.buildImageHyperlinkAnchor(img, run.hyperlink, 'inline-block');
-  }
-
-  /**
-   * Renders a FieldAnnotationRun as an inline "pill" element matching super-editor's visual appearance.
-   *
-   * Field annotations are styled inline elements that display form fields with:
-   * - Outer span with border, border-radius, padding, and background color
-   * - Inner span containing the displayLabel or type-specific content (image, link, etc.)
-   *
-   * @param run - The FieldAnnotationRun to render containing field configuration and styling
-   * @returns HTMLElement (span) or null if document is not available
-   *
-   * @example
-   * ```typescript
-   * // Text variant
-   * renderFieldAnnotationRun({ kind: 'fieldAnnotation', variant: 'text', displayLabel: 'Full Name', fieldColor: '#980043' })
-   * // Returns: <span class="annotation" style="border: 2px solid #b015b3; ..."><span class="annotation-content">Full Name</span></span>
-   *
-   * // Image variant with imageSrc
-   * renderFieldAnnotationRun({ kind: 'fieldAnnotation', variant: 'image', displayLabel: 'Photo', imageSrc: 'data:image/png;...' })
-   * // Returns: <span class="annotation"><span class="annotation-content"><img src="..." /></span></span>
-   *
-   * // Link variant
-   * renderFieldAnnotationRun({ kind: 'fieldAnnotation', variant: 'link', displayLabel: 'Website', linkUrl: 'https://example.com' })
-   * // Returns: <span class="annotation"><span class="annotation-content"><a href="...">https://example.com</a></span></span>
-   * ```
-   */
-  private renderFieldAnnotationRun(run: FieldAnnotationRun): HTMLElement | null {
-    if (!this.doc) {
-      return null;
-    }
-
-    // Handle hidden fields
-    if (run.hidden) {
-      const hidden = this.doc.createElement('span');
-      hidden.style.display = 'none';
-      if (run.pmStart != null) hidden.dataset.pmStart = String(run.pmStart);
-      if (run.pmEnd != null) hidden.dataset.pmEnd = String(run.pmEnd);
-      hidden.dataset.layoutEpoch = String(this.layoutEpoch);
-      return hidden;
-    }
-
-    // Default styling values (matching super-editor's FieldAnnotationView)
-    const defaultBorderColor = '#b015b3';
-    const defaultFieldColor = '#980043';
-
-    // Create outer annotation wrapper
-    const annotation = this.doc.createElement('span');
-    annotation.classList.add(DOM_CLASS_NAMES.ANNOTATION);
-    annotation.setAttribute('aria-label', 'Field annotation');
-
-    // Apply pill styling (unless highlighted is explicitly false)
-    const showHighlight = run.highlighted !== false;
-    if (showHighlight) {
-      const borderColor = run.borderColor || defaultBorderColor;
-      annotation.style.border = `2px solid ${borderColor}`;
-      annotation.style.borderRadius = '2px';
-      annotation.style.padding = '1px 2px';
-      annotation.style.boxSizing = 'border-box';
-
-      // Apply background color with alpha
-      const fieldColor = run.fieldColor || defaultFieldColor;
-      // Add alpha to make it semi-transparent (matching super-editor's behavior)
-      const bgColor = fieldColor.length === 7 ? `${fieldColor}33` : fieldColor;
-      // textHighlight takes precedence over fieldColor
-      if (run.textHighlight) {
-        annotation.style.backgroundColor = run.textHighlight;
-      } else {
-        annotation.style.backgroundColor = bgColor;
-      }
-    }
-
-    // Apply visibility
-    if (run.visibility === 'hidden') {
-      annotation.style.visibility = 'hidden';
-    }
-
-    // Apply explicit size if present
-    if (run.size) {
-      if (run.size.width) {
-        const requiresImage = run.variant === 'image' || run.variant === 'signature';
-        if (!requiresImage || run.imageSrc) {
-          annotation.style.width = `${run.size.width}px`;
-          annotation.style.display = 'inline-block';
-          annotation.style.overflow = 'hidden';
-        }
-      }
-      if (run.size.height && run.variant !== 'html') {
-        const requiresImage = run.variant === 'image' || run.variant === 'signature';
-        if (!requiresImage || run.imageSrc) {
-          annotation.style.height = `${run.size.height}px`;
-        }
-      }
-    }
-
-    // Apply typography to the annotation element.
-    // Always set a font-size so the annotation never inherits fontSize: 0 from
-    // the line container (which zeroes it to eliminate the CSS strut). When the
-    // run has no explicit fontSize, fall back to BROWSER_DEFAULT_FONT_SIZE (the
-    // browser default that was previously inherited before the strut fix).
-    if (run.fontFamily) {
-      annotation.style.fontFamily = run.fontFamily;
-    }
-    {
-      const fontSize = run.fontSize
-        ? typeof run.fontSize === 'number'
-          ? `${run.fontSize}pt`
-          : run.fontSize
-        : BROWSER_DEFAULT_FONT_SIZE;
-      annotation.style.fontSize = fontSize;
-    }
-    if (run.textColor) {
-      annotation.style.color = run.textColor;
-    }
-    if (run.bold) {
-      annotation.style.fontWeight = 'bold';
-    }
-    if (run.italic) {
-      annotation.style.fontStyle = 'italic';
-    }
-    if (run.underline) {
-      annotation.style.textDecoration = 'underline';
-    }
-
-    // Apply z-index for proper layering
-    annotation.style.zIndex = '1';
-
-    // Create inner content wrapper
-    const content = this.doc.createElement('span');
-    content.classList.add(DOM_CLASS_NAMES.ANNOTATION_CONTENT);
-    content.style.pointerEvents = 'none';
-    content.setAttribute('contenteditable', 'false');
-
-    // Render type-specific content
-    switch (run.variant) {
-      case 'image':
-      case 'signature': {
-        if (run.imageSrc) {
-          const img = this.doc.createElement('img');
-          // SECURITY: Validate data URLs
-          const isDataUrl = run.imageSrc.startsWith('data:');
-          if (isDataUrl) {
-            if (run.imageSrc.length <= MAX_DATA_URL_LENGTH && VALID_IMAGE_DATA_URL.test(run.imageSrc)) {
-              img.src = run.imageSrc;
-            } else {
-              // Invalid data URL - fall back to displayLabel
-              content.textContent = run.displayLabel;
-              break;
-            }
-          } else {
-            const sanitized = sanitizeHref(run.imageSrc);
-            if (sanitized) {
-              img.src = sanitized.href;
-            } else {
-              content.textContent = run.displayLabel;
-              break;
-            }
-          }
-          img.alt = run.displayLabel;
-          img.style.height = 'auto';
-          img.style.maxWidth = '100%';
-          img.style.pointerEvents = 'none';
-          img.style.verticalAlign = 'middle';
-          if (run.variant === 'signature') {
-            img.style.maxHeight = '28px';
-          }
-          content.appendChild(img);
-          annotation.style.display = 'inline-block';
-          content.style.display = 'inline-block';
-          // Prevent line-height inheritance from the line container from breaking image layout.
-          annotation.style.lineHeight = 'normal';
-          content.style.lineHeight = 'normal';
-        } else {
-          content.textContent = run.displayLabel || (run.variant === 'signature' ? 'Signature' : '');
-        }
-        break;
-      }
-
-      case 'link': {
-        if (run.linkUrl) {
-          const link = this.doc.createElement('a');
-          const sanitized = sanitizeHref(run.linkUrl);
-          if (sanitized) {
-            link.href = sanitized.href;
-            link.target = '_blank';
-            link.rel = 'noopener noreferrer';
-            link.textContent = run.linkUrl;
-            link.style.textDecoration = 'none';
-            content.style.pointerEvents = 'all';
-            content.appendChild(link);
-          } else {
-            content.textContent = run.displayLabel;
-          }
-        } else {
-          content.textContent = run.displayLabel;
-        }
-        break;
-      }
-
-      case 'html': {
-        if (run.rawHtml && typeof run.rawHtml === 'string') {
-          // Note: rawHtml is expected to be sanitized upstream.
-          content.innerHTML = run.rawHtml.trim();
-          annotation.style.display = 'inline-block';
-          content.style.display = 'inline-block';
-          // Prevent line-height inheritance from the line container from affecting HTML layout.
-          annotation.style.lineHeight = 'normal';
-          content.style.lineHeight = 'normal';
-        } else {
-          content.textContent = run.displayLabel;
-        }
-        break;
-      }
-
-      case 'text':
-      case 'checkbox':
-      default: {
-        content.textContent = run.displayLabel;
-        break;
-      }
-    }
-
-    annotation.appendChild(content);
-
-    // Apply data attributes for field tracking
-    annotation.dataset.type = run.variant;
-    annotation.dataset.displayLabel = run.displayLabel;
-    if (run.fieldId) {
-      annotation.dataset.fieldId = run.fieldId;
-    }
-    if (run.fieldType) {
-      annotation.dataset.fieldType = run.fieldType;
-    }
-
-    // Assert PM positions are present for cursor fallback
-    assertPmPositions(run, 'field annotation run');
-
-    // Apply PM position tracking
-    if (run.pmStart != null) {
-      annotation.dataset.pmStart = String(run.pmStart);
-    }
-    if (run.pmEnd != null) {
-      annotation.dataset.pmEnd = String(run.pmEnd);
-    }
-    annotation.dataset.layoutEpoch = String(this.layoutEpoch);
-
-    // Apply SDT metadata
-    this.applySdtDataset(annotation, run.sdt);
-
-    return annotation;
-  }
-
-  /**
-   * Renders a single line of a paragraph block.
-   *
-   * @param block - The paragraph block containing the line
-   * @param line - The line measurement data
-   * @param context - Rendering context with fragment information
-   * @param availableWidthOverride - Optional override for available width used in justification calculations
-   * @param lineIndex - Optional zero-based index of the line within the fragment
-   * @param skipJustify - When true, prevents justification even if alignment is 'justify'
-   * @param preExpandedRuns - Pre-computed result of expandRunsForInlineNewlines; pass when rendering multiple lines of the same paragraph to avoid recomputing per line
-   * @param resolvedListTextStartPx - Optional canonical text-start override for list first lines
-   * @param indentOffsetOverride - When defined, used instead of re-deriving indentOffset from block attrs in the segment positioning path
-   * @param paragraphMarkLeftOffsetOverride - Optional text-start override for positioning presentation-only paragraph marks
-   * @returns The rendered line element
-   */
   private renderLine(
     block: ParagraphBlock,
     line: Line,
@@ -5701,780 +4186,43 @@ export class DomPainter {
       throw new Error('DomPainter: document is not available');
     }
 
-    const expandedBlock = { ...block, runs: preExpandedRuns ?? expandRunsForInlineNewlines(block.runs) };
-    const lineRange = computeLinePmRange(expandedBlock, line);
-    let runsForLine = sliceRunsForLine(expandedBlock, line);
-
-    const el = this.doc.createElement('div');
-    el.classList.add(CLASS_NAMES.line);
-    applyStyles(el, lineStyles(line.lineHeight));
-    el.dataset.layoutEpoch = String(this.layoutEpoch);
-    const paragraphAttrs = (block.attrs as ParagraphAttrs | undefined) ?? {};
-    const styleId = paragraphAttrs.styleId;
-    if (styleId) {
-      el.setAttribute('styleid', styleId);
-    }
-    const pAttrs = block.attrs as ParagraphAttrs | undefined;
-    const isRtl = applyRtlStyles(el, pAttrs);
-
-    if (lineRange.pmStart != null) {
-      el.dataset.pmStart = String(lineRange.pmStart);
-    }
-    if (lineRange.pmEnd != null) {
-      el.dataset.pmEnd = String(lineRange.pmEnd);
-    }
-    const trackedConfig = this.resolveTrackedChangesConfig(block);
-
-    // Preserve PM positions for DOM caret mapping on empty lines.
-    if (runsForLine.length === 0) {
-      const span = this.doc.createElement('span');
-      span.classList.add('superdoc-empty-run');
-      if (lineRange.pmStart != null) {
-        span.dataset.pmStart = String(lineRange.pmStart);
-      }
-      if (lineRange.pmEnd != null) {
-        span.dataset.pmEnd = String(lineRange.pmEnd);
-      }
-      // Restore font-size so the &nbsp; remains a visible caret target
-      // (the line container sets fontSize: 0 to eliminate the CSS strut).
-      span.style.fontSize = `${line.lineHeight}px`;
-      span.innerHTML = '&nbsp;';
-      el.appendChild(span);
-    }
-
-    // Render tab leaders (absolute positioned overlays)
-    if (line.leaders && line.leaders.length > 0) {
-      line.leaders.forEach((ld) => {
-        const leaderEl = this.doc!.createElement('div');
-        leaderEl.classList.add('superdoc-leader');
-        leaderEl.setAttribute('data-style', ld.style);
-        leaderEl.style.position = 'absolute';
-        leaderEl.style.left = `${ld.from}px`;
-        leaderEl.style.width = `${Math.max(0, ld.to - ld.from)}px`;
-        // Align leaders closer to the text baseline using measured descent
-        const baselineOffset = Math.max(1, Math.round(Math.max(1, line.descent * 0.5)));
-        leaderEl.style.bottom = `${baselineOffset}px`;
-        leaderEl.style.height = ld.style === 'heavy' ? '2px' : '1px';
-        leaderEl.style.pointerEvents = 'none';
-        leaderEl.style.zIndex = '0'; // Same layer as line, text will be z-index: 1
-
-        // Map leader styles to CSS
-        if (ld.style === 'dot' || ld.style === 'middleDot') {
-          leaderEl.style.borderBottom = '1px dotted currentColor';
-        } else if (ld.style === 'hyphen') {
-          leaderEl.style.borderBottom = '1px dashed currentColor';
-        } else if (ld.style === 'underscore') {
-          leaderEl.style.borderBottom = '1px solid currentColor';
-        } else if (ld.style === 'heavy') {
-          leaderEl.style.borderBottom = '2px solid currentColor';
-        }
-
-        el.appendChild(leaderEl);
-      });
-    }
-
-    // Render bar tabs (vertical hairlines)
-    if (line.bars && line.bars.length > 0) {
-      line.bars.forEach((bar) => {
-        const barEl = this.doc!.createElement('div');
-        barEl.classList.add('superdoc-tab-bar');
-        barEl.style.position = 'absolute';
-        barEl.style.left = `${bar.x}px`;
-        barEl.style.top = '0px';
-        barEl.style.bottom = '0px';
-        barEl.style.width = '1px';
-        barEl.style.background = 'currentColor';
-        barEl.style.opacity = '0.6';
-        barEl.style.pointerEvents = 'none';
-        el.appendChild(barEl);
-      });
-    }
-
-    // Check if any segments have explicit X positioning (from tab stops)
-    const hasExplicitPositioning = line.segments?.some((seg) => seg.x !== undefined);
-    const explicitPositionedSegmentCount = line.segments?.filter((seg) => seg.x !== undefined).length ?? 0;
-    const hasMultipleExplicitPositionedSegments = explicitPositionedSegmentCount > 1;
-    const availableWidth = availableWidthOverride ?? line.maxWidth ?? line.width;
-
-    const justifyShouldApply = shouldApplyJustify({
-      alignment: (block as ParagraphBlock).attrs?.alignment,
-      hasExplicitPositioning: hasExplicitPositioning ?? false,
-      hasExplicitTabStops: line.hasExplicitTabStops === true,
-      // Caller already folds last-line + trailing lineBreak behavior into skipJustify.
-      isLastLineOfParagraph: false,
-      paragraphEndsWithLineBreak: false,
-      skipJustifyOverride: skipJustify || hasMultipleExplicitPositionedSegments,
-    });
-
-    const countSpaces = (text: string): number => {
-      let count = 0;
-      for (let i = 0; i < text.length; i += 1) {
-        if (SPACE_CHARS.has(text[i])) count += 1;
-      }
-      return count;
-    };
-
-    if (justifyShouldApply) {
-      // The measurer trims wrap-point trailing spaces from line ranges, but slicing can still
-      // produce whitespace-only runs at style boundaries. These runs are especially problematic
-      // for justify because `word-spacing` behavior is inconsistent on pure-whitespace spans.
-      //
-      // Normalize by merging whitespace-only slices into adjacent runs with identical styling.
-      const stableDataAttrs = (attrs: Record<string, string> | undefined): Record<string, string> | undefined => {
-        if (!attrs) return undefined;
-        const keys = Object.keys(attrs).sort();
-        const out: Record<string, string> = {};
-        keys.forEach((key) => {
-          out[key] = attrs[key]!;
-        });
-        return out;
-      };
-
-      const mergeSignature = (run: TextRun): string =>
-        JSON.stringify({
-          kind: run.kind ?? 'text',
-          fontFamily: run.fontFamily,
-          fontSize: run.fontSize,
-          bold: run.bold ?? false,
-          italic: run.italic ?? false,
-          letterSpacing: run.letterSpacing ?? null,
-          color: run.color ?? null,
-          underline: run.underline ?? null,
-          strike: run.strike ?? false,
-          highlight: run.highlight ?? null,
-          textTransform: run.textTransform ?? null,
-          token: run.token ?? null,
-          pageRefMetadata: run.pageRefMetadata ?? null,
-          trackedChange: run.trackedChange ?? null,
-          sdt: run.sdt ?? null,
-          link: run.link ?? null,
-          comments: run.comments ?? null,
-          dataAttrs: stableDataAttrs(run.dataAttrs) ?? null,
-          bidi: run.bidi ?? null,
-        });
-
-      const isWhitespaceOnly = (text: string): boolean => {
-        if (text.length === 0) return false;
-        for (let i = 0; i < text.length; i += 1) {
-          if (!SPACE_CHARS.has(text[i])) return false;
-        }
-        return true;
-      };
-
-      const cloneTextRun = (run: TextRun): TextRun => ({
-        ...(run as TextRun),
-        comments: run.comments ? [...run.comments] : undefined,
-        dataAttrs: run.dataAttrs ? { ...run.dataAttrs } : undefined,
-        underline: run.underline ? { ...run.underline } : undefined,
-        pageRefMetadata: run.pageRefMetadata ? { ...run.pageRefMetadata } : undefined,
-      });
-
-      const normalized: Run[] = runsForLine.map((run) => {
-        if ((run.kind !== 'text' && run.kind !== undefined) || !('text' in run)) return run;
-        return cloneTextRun(run as TextRun);
-      });
-
-      const merged: Run[] = [];
-      for (let i = 0; i < normalized.length; i += 1) {
-        const run = normalized[i]!;
-        if ((run.kind !== 'text' && run.kind !== undefined) || !('text' in run)) {
-          merged.push(run);
-          continue;
-        }
-
-        const textRun = run as TextRun;
-        if (!isWhitespaceOnly(textRun.text ?? '')) {
-          merged.push(textRun);
-          continue;
-        }
-
-        const prev = merged[merged.length - 1];
-        if (prev && (prev.kind === 'text' || prev.kind === undefined) && 'text' in prev) {
-          const prevTextRun = prev as TextRun;
-          if (mergeSignature(prevTextRun) === mergeSignature(textRun)) {
-            const extra = textRun.text ?? '';
-            prevTextRun.text = (prevTextRun.text ?? '') + extra;
-            if (prevTextRun.pmStart != null) {
-              prevTextRun.pmEnd = prevTextRun.pmStart + prevTextRun.text.length;
-            } else if (prevTextRun.pmEnd != null) {
-              prevTextRun.pmEnd = prevTextRun.pmEnd + extra.length;
-            }
-            continue;
-          }
-        }
-
-        const next = normalized[i + 1];
-        if (next && (next.kind === 'text' || next.kind === undefined) && 'text' in next) {
-          const nextTextRun = next as TextRun;
-          if (mergeSignature(nextTextRun) === mergeSignature(textRun)) {
-            const extra = textRun.text ?? '';
-            nextTextRun.text = extra + (nextTextRun.text ?? '');
-            if (textRun.pmStart != null) {
-              nextTextRun.pmStart = textRun.pmStart;
-            } else if (nextTextRun.pmStart != null) {
-              nextTextRun.pmStart = nextTextRun.pmStart - extra.length;
-            }
-            if (nextTextRun.pmStart != null && nextTextRun.pmEnd == null) {
-              nextTextRun.pmEnd = nextTextRun.pmStart + nextTextRun.text.length;
-            }
-            continue;
-          }
-        }
-
-        merged.push(textRun);
-      }
-
-      runsForLine = merged;
-
-      // Suppress trailing wrap-point spaces on justified lines. With `white-space: pre`, they would
-      // otherwise consume width and be stretched by word-spacing, producing a ragged visible edge.
-      // Preserve intentionally space-only lines (rare but supported).
-      const hasNonSpaceText = runsForLine.some(
-        (run) => (run.kind === 'text' || run.kind === undefined) && 'text' in run && (run.text ?? '').trim().length > 0,
-      );
-      if (hasNonSpaceText) {
-        for (let i = runsForLine.length - 1; i >= 0; i -= 1) {
-          const run = runsForLine[i];
-          if ((run.kind !== 'text' && run.kind !== undefined) || !('text' in run)) continue;
-          const text = run.text ?? '';
-          let trimCount = 0;
-          for (let j = text.length - 1; j >= 0 && text[j] === ' '; j -= 1) {
-            trimCount += 1;
-          }
-          if (trimCount === 0) break;
-
-          const nextText = text.slice(0, Math.max(0, text.length - trimCount));
-          if (nextText.length === 0) {
-            runsForLine.splice(i, 1);
-            continue;
-          }
-          (run as TextRun).text = nextText;
-          if ((run as TextRun).pmEnd != null) {
-            (run as TextRun).pmEnd = (run as TextRun).pmEnd! - trimCount;
-          }
-          break;
-        }
-      }
-    }
-
-    const spaceCount =
-      line.spaceCount ??
-      runsForLine.reduce((sum, run) => {
-        if ((run.kind !== 'text' && run.kind !== undefined) || !('text' in run) || run.text == null) return sum;
-        return sum + countSpaces(run.text);
-      }, 0);
-    const lineWidth = line.naturalWidth ?? line.width;
-    const spacingPerSpace = calculateJustifySpacing({
-      lineWidth,
-      availableWidth,
-      spaceCount,
-      shouldJustify: justifyShouldApply,
-    });
-    const resolveLineIndentOffset = (): number => {
-      if (indentOffsetOverride != null) {
-        return indentOffsetOverride;
-      }
-
-      const paraIndent = (block.attrs as ParagraphAttrs | undefined)?.indent;
-      const indentLeft = paraIndent?.left ?? 0;
-      const firstLine = paraIndent?.firstLine ?? 0;
-      const hanging = paraIndent?.hanging ?? 0;
-      const isFirstLineOfPara = lineIndex === 0 || lineIndex === undefined;
-      const firstLineOffsetForCumX = isFirstLineOfPara ? firstLine - hanging : 0;
-      const wordLayoutValue = (block.attrs as ParagraphAttrs | undefined)?.wordLayout;
-      const wordLayout = isMinimalWordLayout(wordLayoutValue) ? wordLayoutValue : undefined;
-      const isListParagraph = Boolean(wordLayout?.marker);
-      const fallbackListTextStartPx =
-        typeof wordLayout?.marker?.textStartX === 'number' && Number.isFinite(wordLayout.marker.textStartX)
-          ? wordLayout.marker.textStartX
-          : typeof wordLayout?.textStartPx === 'number' && Number.isFinite(wordLayout.textStartPx)
-            ? wordLayout.textStartPx
-            : undefined;
-      const listIndentOffset = isFirstLineOfPara
-        ? (resolvedListTextStartPx ?? fallbackListTextStartPx ?? indentLeft)
-        : indentLeft;
-
-      return isListParagraph ? listIndentOffset : indentLeft + firstLineOffsetForCumX;
-    };
-    const lineTextStartOffsetPx =
-      paragraphMarkLeftOffsetOverride != null ? paragraphMarkLeftOffsetOverride : resolveLineIndentOffset();
-    const paragraphMarkLeftOffsetPx = lineTextStartOffsetPx;
-
-    if (spacingPerSpace !== 0) {
-      // Each rendered line is its own block; relying on text-align-last is brittle, so we use word-spacing.
-      el.style.wordSpacing = `${spacingPerSpace}px`;
-    }
-
-    if (shouldUseSegmentPositioning(hasExplicitPositioning ?? false, Boolean(line.segments), isRtl)) {
-      // Use segment-based rendering with absolute positioning for tab-aligned text.
-      // shouldUseSegmentPositioning returns false for RTL because the layout engine
-      // computes tab positions in LTR order; RTL lines fall through to inline-flow
-      // rendering where dir="rtl" lets the browser handle tab positioning.
-      //
-      // The segment x positions from layout are relative to the content area (left margin = 0).
-      // We need to add the paragraph indent to ALL positions (both explicit and calculated).
-      // Segment x positions and paragraph marks both need the visual text start,
-      // including list marker/suffix space when the resolved layout provides it.
-      const indentOffset = lineTextStartOffsetPx;
-      let cumulativeX = 0; // Start at 0, we'll add indentOffset when positioning
-
-      const segments = line.segments!;
-      const segmentsByRun = new Map<number, LineSegment[]>();
-      segments.forEach((segment) => {
-        const list = segmentsByRun.get(segment.runIndex);
-        if (list) {
-          list.push(segment);
-        } else {
-          segmentsByRun.set(segment.runIndex, [segment]);
-        }
-      });
-
-      /**
-       * Finds the immediate next segment carrying tab geometry after a given run index.
-       * This handles tab-aligned text and compensated tab paint geometry.
-       *
-       * WHY ONLY THE IMMEDIATE NEXT RUN:
-       * When rendering a tab, we need to know where the content IMMEDIATELY after this tab begins
-       * to correctly size the tab element. We don't look beyond the immediate next run because:
-       * 1. Each tab is independent and should only consider its directly adjacent content
-       * 2. Looking further ahead would incorrectly span multiple tabs or unrelated runs
-       * 3. If there's another tab between this tab and some content, that intermediate tab is
-       *    responsible for its own layout - we shouldn't reach across it
-       *
-       * For example, given: "Text[TAB1]Content[TAB2]MoreContent"
-       * - When sizing TAB1, we only check "Content" (immediate next run)
-       * - We don't check "MoreContent" because TAB2 is in between
-       * - TAB2 will independently check "MoreContent" when it's rendered
-       *
-       * @param fromRunIndex - The run index to search after
-       * @returns The immediate next tab-positioned segment, or undefined if not found or not immediate
-       */
-      const findImmediateNextSegment = (fromRunIndex: number): LineSegment | undefined => {
-        // Only check the immediate next run - don't skip over other tabs
-        const nextRunIdx = fromRunIndex + 1;
-        if (nextRunIdx <= line.toRun) {
-          const nextSegments = segmentsByRun.get(nextRunIdx);
-          if (nextSegments && nextSegments.length > 0) {
-            const firstSegment = nextSegments[0];
-            // Return only the first segment; later segments in the same run are
-            // not immediately adjacent to this tab.
-            return firstSegment.x !== undefined || firstSegment.precedingTabEndX !== undefined
-              ? firstSegment
-              : undefined;
-          }
-        }
-        return undefined;
-      };
-
-      // Inline SDT wrapping for geometry path (absolute-positioned elements).
-      // Same concept as the run-based path's SDT wrapper, but here elements use
-      // position:absolute so the wrapper itself must be absolutely positioned to
-      // span from the leftmost to rightmost child element.
-      let geoSdtWrapper: HTMLElement | null = null;
-      let geoSdtId: string | null = null;
-      let geoSdtWrapperLeft = 0;
-      let geoSdtMaxRight = 0;
-
-      const closeGeoSdtWrapper = () => {
-        if (geoSdtWrapper) {
-          geoSdtWrapper.style.width = `${geoSdtMaxRight - geoSdtWrapperLeft}px`;
-          el.appendChild(geoSdtWrapper);
-          geoSdtWrapper = null;
-          geoSdtId = null;
-        }
-      };
-
-      /**
-       * Append an element to the line, routing through an inline SDT wrapper
-       * when the run has inline structuredContent metadata.
-       */
-      const appendToLineGeo = (elem: HTMLElement, runForSdt: Run, elemLeftPx: number, elemWidthPx: number) => {
-        const resolved = this.resolveRunSdtId(runForSdt);
-        const thisRunSdtId = resolved?.sdtId ?? null;
-
-        if (thisRunSdtId !== geoSdtId) {
-          closeGeoSdtWrapper();
-        }
-
-        if (resolved && this.doc) {
-          if (!geoSdtWrapper) {
-            geoSdtWrapper = this.createInlineSdtWrapper(resolved.sdt);
-            geoSdtId = thisRunSdtId;
-            geoSdtWrapperLeft = elemLeftPx;
-            geoSdtMaxRight = elemLeftPx;
-            geoSdtWrapper.style.position = 'absolute';
-            geoSdtWrapper.style.left = `${elemLeftPx}px`;
-            geoSdtWrapper.style.top = '0px';
-            geoSdtWrapper.style.height = `${line.lineHeight}px`;
-          }
-          this.syncInlineSdtWrapperTypography(geoSdtWrapper, runForSdt);
-          elem.style.left = `${elemLeftPx - geoSdtWrapperLeft}px`;
-          geoSdtMaxRight = Math.max(geoSdtMaxRight, elemLeftPx + elemWidthPx);
-          this.expandSdtWrapperPmRange(geoSdtWrapper, (runForSdt as TextRun).pmStart, (runForSdt as TextRun).pmEnd);
-          geoSdtWrapper.appendChild(elem);
-        } else {
-          el.appendChild(elem);
-        }
-      };
-
-      for (let runIndex = line.fromRun; runIndex <= line.toRun; runIndex += 1) {
-        const baseRun = block.runs[runIndex];
-        if (!baseRun) continue;
-
-        if (baseRun.kind === 'tab') {
-          // Find where the immediate next content begins (if it's right after this tab)
-          const immediateNextSegment = findImmediateNextSegment(runIndex);
-          const tabStartX = cumulativeX;
-
-          // The tab should span from where previous content ended to where next content begins.
-          // If layout supplied a tab-end boundary for the next segment, prefer it.
-          // Otherwise, use the next segment's explicit X (from tab alignment) or the
-          // tab's measured width.
-          const measuredTabEndX = tabStartX + (baseRun.width ?? 0);
-          const tabEndX = immediateNextSegment?.precedingTabEndX ?? immediateNextSegment?.x ?? measuredTabEndX;
-          const actualTabWidth = tabEndX - tabStartX;
-
-          const tabEl = this.doc!.createElement('span');
-          tabEl.style.position = 'absolute';
-          tabEl.style.left = `${tabStartX + indentOffset}px`;
-          tabEl.style.top = '0px';
-          tabEl.style.width = `${actualTabWidth}px`;
-          tabEl.style.height = `${line.lineHeight}px`;
-          tabEl.style.display = 'inline-block';
-          tabEl.style.pointerEvents = 'none';
-          tabEl.style.zIndex = '1';
-
-          // Apply underline styling to tab if present (common in signature lines)
-          // TabRun can have RunMarks properties like underline, bold, etc.
-          //
-          // Signature line use case: In documents with signature lines, tabs are often used
-          // to create underlined blank spaces where signatures should go. The underline mark
-          // is inherited from a parent node (e.g., a paragraph with underline formatting) and
-          // applied to the tab, creating a visible underline even though the tab itself has
-          // no visible text content.
-          if (baseRun.underline) {
-            const underlineStyle = baseRun.underline.style ?? 'single';
-            // We must use an explicit color instead of currentColor because tab content is
-            // invisible (no text). If we used currentColor, the underline would inherit the
-            // text color, which might be transparent or the same as the background, making
-            // the underline invisible. Using an explicit color (defaulting to black) ensures
-            // the underline is always visible for signature lines.
-            const underlineColor = baseRun.underline.color ?? '#000000';
-            const borderStyle = underlineStyle === 'double' ? 'double' : 'solid';
-            tabEl.style.borderBottom = `1px ${borderStyle} ${underlineColor}`;
-          } else {
-            tabEl.style.visibility = 'hidden';
-          }
-
-          if (styleId) {
-            tabEl.setAttribute('styleid', styleId);
-          }
-          if (baseRun.pmStart != null) tabEl.dataset.pmStart = String(baseRun.pmStart);
-          if (baseRun.pmEnd != null) tabEl.dataset.pmEnd = String(baseRun.pmEnd);
-          tabEl.dataset.layoutEpoch = String(this.layoutEpoch);
-          appendToLineGeo(tabEl, baseRun, tabStartX + indentOffset, actualTabWidth);
-
-          // Update cumulativeX to where the next content begins
-          // This ensures proper positioning for subsequent elements
-          cumulativeX = tabEndX;
-          continue;
-        }
-
-        // Handle ImageRun - render as-is (no slicing needed, atomic unit)
-        if (this.isImageRun(baseRun)) {
-          const elem = this.renderRun(baseRun, context, trackedConfig);
-          if (elem) {
-            if (styleId) {
-              elem.setAttribute('styleid', styleId);
-            }
-            // Position image using explicit segment X when available; fallback to cumulative flow
-            // Add indentOffset to position content at the correct paragraph indent.
-            const runSegments = segmentsByRun.get(runIndex);
-            const baseSegX = runSegments && runSegments[0]?.x !== undefined ? runSegments[0].x : cumulativeX;
-            const segX = baseSegX + indentOffset;
-            // LineSegment.width is required by contract; producer (measuring-dom) always emits it.
-            // No paint-time DOM measurement (SD-2957).
-            const segWidth = runSegments?.[0]?.width ?? 0;
-            elem.style.position = 'absolute';
-            elem.style.left = `${segX}px`;
-            appendToLineGeo(elem, baseRun, segX, segWidth);
-            cumulativeX = baseSegX + segWidth;
-          }
-          continue;
-        }
-
-        // Handle LineBreakRun - line breaks are handled by line creation, skip here
-        if (this.isLineBreakRun(baseRun)) {
-          continue;
-        }
-
-        // Handle BreakRun - breaks are handled by line creation, skip here
-        if (this.isBreakRun(baseRun)) {
-          continue;
-        }
-
-        // Handle FieldAnnotationRun - render as-is (no slicing needed, atomic unit like images)
-        if (this.isFieldAnnotationRun(baseRun)) {
-          const elem = this.renderRun(baseRun, context, trackedConfig);
-          if (elem) {
-            if (styleId) {
-              elem.setAttribute('styleid', styleId);
-            }
-            // Position using explicit segment X when available; fallback to cumulative flow
-            // Add indentOffset to position content at the correct paragraph indent.
-            const runSegments = segmentsByRun.get(runIndex);
-            const baseSegX = runSegments && runSegments[0]?.x !== undefined ? runSegments[0].x : cumulativeX;
-            const segX = baseSegX + indentOffset;
-            const segWidth = (runSegments && runSegments[0]?.width !== undefined ? runSegments[0].width : 0) ?? 0;
-            elem.style.position = 'absolute';
-            elem.style.left = `${segX}px`;
-            appendToLineGeo(elem, baseRun, segX, segWidth);
-            cumulativeX = baseSegX + segWidth;
-          }
-          continue;
-        }
-
-        // Handle MathRun - render as-is (atomic unit like images)
-        if (this.isMathRun(baseRun)) {
-          const elem = this.renderRun(baseRun, context, trackedConfig);
-          if (elem) {
-            if (styleId) {
-              elem.setAttribute('styleid', styleId);
-            }
-            const runSegments = segmentsByRun.get(runIndex);
-            const baseSegX = runSegments && runSegments[0]?.x !== undefined ? runSegments[0].x : cumulativeX;
-            const segX = baseSegX + indentOffset;
-            const segWidth =
-              (runSegments && runSegments[0]?.width !== undefined ? runSegments[0].width : baseRun.width) ?? 0;
-            elem.style.position = 'absolute';
-            elem.style.left = `${segX}px`;
-            appendToLineGeo(elem, baseRun, segX, segWidth);
-            cumulativeX = baseSegX + segWidth;
-          }
-          continue;
-        }
-
-        const runSegments = segmentsByRun.get(runIndex);
-        if (!runSegments || runSegments.length === 0) {
-          continue;
-        }
-
-        // At this point, baseRun must be TextRun (has .text property)
-        if (!('text' in baseRun)) {
-          continue;
-        }
-
-        const baseText = baseRun.text ?? '';
-        const runPmStart = baseRun.pmStart ?? null;
-        const fallbackPmEnd =
-          runPmStart != null && baseRun.pmEnd == null ? runPmStart + baseText.length : (baseRun.pmEnd ?? null);
-
-        runSegments.forEach((segment) => {
-          const segmentText = baseText.slice(segment.fromChar, segment.toChar);
-          if (!segmentText) return;
-
-          const pmSliceStart = runPmStart != null ? runPmStart + segment.fromChar : undefined;
-          const pmSliceEnd = runPmStart != null ? runPmStart + segment.toChar : (fallbackPmEnd ?? undefined);
-          const segmentRun: TextRun = {
-            ...(baseRun as TextRun),
-            text: segmentText,
-            pmStart: pmSliceStart,
-            pmEnd: pmSliceEnd,
-          };
-
-          const elem = this.renderRun(segmentRun, context, trackedConfig);
-          if (elem) {
-            if (styleId) {
-              elem.setAttribute('styleid', styleId);
-            }
-            // Determine X position for this segment
-            // Layout positions are relative to content area start (0).
-            // Add indentOffset to position content at the correct paragraph indent.
-            const baseX = segment.x !== undefined ? segment.x : cumulativeX;
-            const xPos = baseX + indentOffset;
-
-            elem.style.position = 'absolute';
-            elem.style.left = `${xPos}px`;
-            appendToLineGeo(elem, segmentRun, xPos, segment.width);
-
-            // Advance cumulative X by the resolved segment width. LineSegment.width is the
-            // sole source of truth — the painter does not measure inline elements (SD-2957).
-            // Use baseX (without indent) to keep cumulativeX relative to content area,
-            // matching how segment.x values are calculated in layout.
-            const width = segment.width;
-            const justifyExtraWidth = spacingPerSpace !== 0 ? spacingPerSpace * countSpaces(segmentText) : 0;
-            const visualWidth = width + justifyExtraWidth;
-            cumulativeX = baseX + visualWidth;
-            // Update SDT wrapper width if actual measured width differs from initial estimate
-            if (geoSdtWrapper) {
-              geoSdtMaxRight = Math.max(geoSdtMaxRight, xPos + visualWidth);
-            }
-          }
-        });
-      }
-      // Close any remaining SDT wrapper at end of geometry rendering
-      closeGeoSdtWrapper();
-    } else {
-      // Use run-based rendering for normal text flow
-      // Track current inline SDT wrapper to group adjacent runs with the same SDT id
-      let currentInlineSdtWrapper: HTMLElement | null = null;
-      let currentInlineSdtId: string | null = null;
-
-      const closeCurrentWrapper = () => {
-        if (currentInlineSdtWrapper) {
-          el.appendChild(currentInlineSdtWrapper);
-          currentInlineSdtWrapper = null;
-          currentInlineSdtId = null;
-        }
-      };
-
-      runsForLine.forEach((run) => {
-        // Check if this run has inline structuredContent SDT
-        const resolved = this.resolveRunSdtId(run);
-        const runSdtId = resolved?.sdtId ?? null;
-
-        // If SDT context changed, close the current wrapper
-        if (runSdtId !== currentInlineSdtId) {
-          closeCurrentWrapper();
-        }
-
-        // Special handling for TabRuns (e.g., signature lines with underlines)
-        let elem: HTMLElement | null = null;
-        if (run.kind === 'tab') {
-          const tabEl = this.doc!.createElement('span');
-          tabEl.classList.add('superdoc-tab');
-
-          // Calculate tab width - use measured width or estimate based on typical tab stop
-          const tabWidth = run.width ?? 48; // Default tab width if not measured
-
-          tabEl.style.display = 'inline-block';
-          tabEl.style.width = `${tabWidth}px`;
-          tabEl.style.height = `${line.lineHeight}px`;
-          tabEl.style.verticalAlign = 'bottom';
-
-          // Apply underline styling if present (common for signature lines)
-          //
-          // Signature line use case: In documents with signature lines, tabs are often used
-          // to create underlined blank spaces where signatures should go. The underline mark
-          // is inherited from a parent node (e.g., a paragraph with underline formatting) and
-          // applied to the tab, creating a visible underline even though the tab itself has
-          // no visible text content.
-          if (run.underline) {
-            const underlineStyle = run.underline.style ?? 'single';
-            // We must use an explicit color instead of currentColor because tab content is
-            // invisible (no text). If we used currentColor, the underline would inherit the
-            // text color, which might be transparent or the same as the background, making
-            // the underline invisible. Using an explicit color (defaulting to black) ensures
-            // the underline is always visible for signature lines.
-            const underlineColor = run.underline.color ?? '#000000';
-            const borderStyle = underlineStyle === 'double' ? 'double' : 'solid';
-            tabEl.style.borderBottom = `1px ${borderStyle} ${underlineColor}`;
-          }
-
-          if (styleId) {
-            tabEl.setAttribute('styleid', styleId);
-          }
-          if (run.pmStart != null) tabEl.dataset.pmStart = String(run.pmStart);
-          if (run.pmEnd != null) tabEl.dataset.pmEnd = String(run.pmEnd);
-          tabEl.dataset.layoutEpoch = String(this.layoutEpoch);
-
-          elem = tabEl;
-        } else {
-          elem = this.renderRun(run, context, trackedConfig);
-        }
-        if (elem) {
-          if (styleId) {
-            elem.setAttribute('styleid', styleId);
-          }
-
-          // If this run has inline SDT, add to or create wrapper
-          if (resolved && this.doc) {
-            if (!currentInlineSdtWrapper) {
-              currentInlineSdtWrapper = this.createInlineSdtWrapper(resolved.sdt);
-              this.syncInlineSdtWrapperTypography(currentInlineSdtWrapper, run);
-              currentInlineSdtId = runSdtId;
-            }
-            // Typography is set when wrapper is created from the first run.
-            // Follow-up (SD-2744): define a deterministic mixed-typography rule.
-            this.expandSdtWrapperPmRange(currentInlineSdtWrapper, run.pmStart, run.pmEnd);
-            currentInlineSdtWrapper.appendChild(elem);
-          } else {
-            el.appendChild(elem);
-          }
-        }
-      });
-
-      // Close any remaining wrapper at end of line
-      closeCurrentWrapper();
-    }
-
-    this.appendFormattingParagraphMark(
-      el,
+    return renderRunLine({
+      block,
       line,
-      expandedBlock.runs,
-      paragraphMarkLeftOffsetPx,
-      availableWidth,
-      hasExplicitPositioning ?? false,
-    );
-
-    // Post-process: Apply tooltip accessibility for any links with pending tooltips
-    // This must happen after elements are in the DOM so aria-describedby can reference siblings
-    const anchors = el.querySelectorAll('a[href]');
-    anchors.forEach((anchor) => {
-      const pendingTooltip = this.pendingTooltips.get(anchor as HTMLElement);
-      if (pendingTooltip) {
-        this.applyTooltipAccessibility(anchor as HTMLAnchorElement, pendingTooltip);
-        this.pendingTooltips.delete(anchor as HTMLElement); // Clean up memory
-      }
+      context,
+      availableWidthOverride,
+      lineIndex,
+      skipJustify,
+      preExpandedRuns,
+      resolvedListTextStartPx,
+      indentOffsetOverride,
+      paragraphMarkLeftOffsetOverride,
+      runContext: this.createRunRenderContext(),
     });
-
-    return el;
   }
 
-  private resolveTrackedChangesConfig(block: ParagraphBlock): TrackedChangesRenderConfig {
-    const attrs = (block.attrs as ParagraphAttrs | undefined) ?? {};
-    const mode = (attrs.trackedChangesMode as TrackedChangesMode | undefined) ?? 'review';
-    const enabled = attrs.trackedChangesEnabled !== false;
-    return { mode, enabled };
-  }
-
-  private applyTrackedChangeDecorations(elem: HTMLElement, run: Run, config: TrackedChangesRenderConfig): void {
-    if (!config.enabled || config.mode === 'off') {
-      return;
+  private createRunRenderContext(): RunRenderContext {
+    if (!this.doc) {
+      throw new Error('DomPainter: document is not available');
     }
 
-    const textRun = run as TextRun;
-    const meta = textRun.trackedChange;
-    if (!meta) {
-      return;
-    }
-
-    const baseClass = TRACK_CHANGE_BASE_CLASS[meta.kind];
-    if (baseClass) {
-      elem.classList.add(baseClass);
-    }
-
-    const modifier = TRACK_CHANGE_MODIFIER_CLASS[meta.kind]?.[config.mode];
-    if (modifier) {
-      elem.classList.add(modifier);
-    }
-
-    elem.dataset.trackChangeId = meta.id;
-    elem.dataset.trackChangeKind = meta.kind;
-    elem.dataset.storyKey = meta.storyKey ?? 'body';
-    if (meta.author) {
-      elem.dataset.trackChangeAuthor = meta.author;
-    }
-    if (meta.authorEmail) {
-      elem.dataset.trackChangeAuthorEmail = meta.authorEmail;
-    }
-    if (meta.authorImage) {
-      elem.dataset.trackChangeAuthorImage = meta.authorImage;
-    }
-    if (meta.date) {
-      elem.dataset.trackChangeDate = meta.date;
-    }
-    // track-change-focused class is applied post-paint by CommentHighlightDecorator (super-editor).
+    return {
+      doc: this.doc,
+      layoutEpoch: this.layoutEpoch,
+      showFormattingMarks: this.showFormattingMarks,
+      pendingTooltips: this.pendingTooltips,
+      getNextLinkId: () => `superdoc-link-${++this.linkIdCounter}`,
+      applySdtDataset: this.applySdtDataset.bind(this),
+      buildImageHyperlinkAnchor: this.buildImageHyperlinkAnchor.bind(
+        this,
+      ) as RunRenderContext['buildImageHyperlinkAnchor'],
+      resolveTrackedChangesConfig,
+      applyTrackedChangeDecorations,
+      resolveRunSdtId: this.resolveRunSdtId.bind(this),
+      createInlineSdtWrapper: this.createInlineSdtWrapper.bind(this),
+      syncInlineSdtWrapperTypography: this.syncInlineSdtWrapperTypography.bind(this),
+      expandSdtWrapperPmRange: this.expandSdtWrapperPmRange.bind(this),
+    };
   }
 
   /**
@@ -7524,105 +5272,6 @@ const deriveBlockVersion = (block: FlowBlock): string => {
   return block.id;
 };
 
-const DEFAULT_SUPERSCRIPT_RAISE_RATIO = 0.33;
-const DEFAULT_SUBSCRIPT_LOWER_RATIO = 0.14;
-
-const hasVerticalPositioning = (run: TextRun): boolean =>
-  normalizeBaselineShift(run.baselineShift) != null || run.vertAlign === 'superscript' || run.vertAlign === 'subscript';
-
-const applyRunVerticalPositioning = (element: HTMLElement, run: TextRun): void => {
-  // Vertically shifted runs should use a tight inline box. If they inherit the
-  // parent line's full line-height, the glyph remains visually low inside an
-  // oversized inline box even when the superscript/subscript offset is correct.
-  if (hasVerticalPositioning(run)) {
-    element.style.lineHeight = '1';
-  }
-
-  const explicitBaselineShift = normalizeBaselineShift(run.baselineShift);
-  if (explicitBaselineShift != null) {
-    element.style.verticalAlign = `${explicitBaselineShift}pt`;
-    return;
-  }
-
-  if (run.vertAlign === 'superscript') {
-    const baseFontSize = resolveBaseFontSizeForVerticalText(run.fontSize, run);
-    element.style.verticalAlign = `${baseFontSize * DEFAULT_SUPERSCRIPT_RAISE_RATIO}px`;
-    return;
-  }
-
-  if (run.vertAlign === 'subscript') {
-    const baseFontSize = resolveBaseFontSizeForVerticalText(run.fontSize, run);
-    element.style.verticalAlign = `${-(baseFontSize * DEFAULT_SUBSCRIPT_LOWER_RATIO)}px`;
-    return;
-  }
-
-  if (run.vertAlign === 'baseline') {
-    element.style.verticalAlign = 'baseline';
-  }
-};
-
-/**
- * Applies run styling properties to a DOM element.
- *
- * @param element - The HTML element to style
- * @param run - The run object containing styling information
- * @param _isLink - Whether this run is part of a hyperlink. Note: This parameter
- *                  is kept for API compatibility but no longer affects behavior -
- *                  inline colors are now applied to all runs (including links) to
- *                  ensure OOXML hyperlink character styles appear correctly.
- */
-const applyRunStyles = (element: HTMLElement, run: Run, _isLink = false): void => {
-  if (
-    run.kind === 'tab' ||
-    run.kind === 'image' ||
-    run.kind === 'lineBreak' ||
-    run.kind === 'break' ||
-    run.kind === 'fieldAnnotation' ||
-    run.kind === 'math'
-  ) {
-    // Tab, image, lineBreak, break, and fieldAnnotation runs don't have text styling properties
-    return;
-  }
-
-  element.style.fontFamily = run.fontFamily;
-  element.style.fontSize = `${run.fontSize}px`;
-  if (run.bold) element.style.fontWeight = 'bold';
-  if (run.italic) element.style.fontStyle = 'italic';
-
-  // Apply inline color even for links so OOXML hyperlink styles appear when CSS is absent
-  if (run.color) element.style.color = run.color;
-
-  if (run.letterSpacing != null) {
-    element.style.letterSpacing = `${run.letterSpacing}px`;
-  }
-  if (run.highlight) {
-    element.style.backgroundColor = run.highlight;
-  }
-  if (run.textTransform) {
-    element.style.textTransform = run.textTransform;
-  }
-
-  // Apply text decorations from the run. Even for links, inline decorations should reflect
-  // the document styling (tests assert underline presence on anchors).
-  const decorations: string[] = [];
-  if (run.underline) {
-    decorations.push('underline');
-    const u = run.underline;
-    element.style.textDecorationStyle = u.style && u.style !== 'single' ? u.style : 'solid';
-    if (u.color) {
-      element.style.textDecorationColor = u.color;
-    }
-  }
-  if (run.strike) {
-    decorations.push('line-through');
-  }
-  if (decorations.length > 0) {
-    element.style.textDecorationLine = decorations.join(' ');
-  }
-
-  applyRunVerticalPositioning(element, run);
-};
-
 const CLIP_PATH_PREFIXES = ['inset(', 'polygon(', 'circle(', 'ellipse(', 'path(', 'rect('];
 
 const readClipPathValue = (value: unknown): string => {
@@ -7646,74 +5295,10 @@ const resolveBlockClipPath = (block: unknown): string => {
   return readClipPathValue(record.clipPath) || resolveClipPathFromAttrs(record.attrs);
 };
 
-/**
- * Applies data-* attributes from a text run to a DOM element.
- * Validates attribute names and safely sets them on the element.
- * Invalid or unsafe attributes are skipped with development-mode logging.
- *
- * @param element - The HTML element to apply attributes to
- * @param dataAttrs - Record of data-* attribute key-value pairs from the text run
- *
- * @example
- * ```typescript
- * const span = document.createElement('span');
- * applyRunDataAttributes(span, { 'data-id': '123', 'data-name': 'test' });
- * // span now has: <span data-id="123" data-name="test"></span>
- * ```
- */
-export const applyRunDataAttributes = (element: HTMLElement, dataAttrs?: Record<string, string>): void => {
-  if (!dataAttrs) return;
-  Object.entries(dataAttrs).forEach(([key, value]) => {
-    if (typeof key !== 'string' || !key.toLowerCase().startsWith('data-')) return;
-    if (typeof value !== 'string') return;
-    try {
-      element.setAttribute(key, value);
-    } catch (error) {
-      if (process.env.NODE_ENV === 'development') {
-        console.warn(`[DomPainter] Failed to set data attribute "${key}":`, error);
-      }
-    }
-  });
-};
-
 const applyStyles = (el: HTMLElement, styles: Partial<CSSStyleDeclaration>): void => {
   Object.entries(styles).forEach(([key, value]) => {
     if (value != null && value !== '' && key in el.style) {
       (el.style as unknown as Record<string, string>)[key] = String(value);
     }
   });
-};
-
-const resolveRunText = (run: Run, context: FragmentRenderContext): string => {
-  const runToken = 'token' in run ? run.token : undefined;
-
-  if (run.kind === 'tab') {
-    return run.text;
-  }
-  if (run.kind === 'image') {
-    // Image runs don't have text content
-    return '';
-  }
-  if (run.kind === 'lineBreak') {
-    // Line break runs don't render text - the measurer creates new lines for them
-    return '';
-  }
-  if (run.kind === 'break') {
-    // Break runs don't render text - the measurer creates new lines for them
-    return '';
-  }
-  if (!('text' in run)) {
-    // Safety check - if run doesn't have text property, return empty string
-    return '';
-  }
-  if (!runToken) {
-    return run.text ?? '';
-  }
-  if (runToken === 'pageNumber') {
-    return context.pageNumberText ?? String(context.pageNumber);
-  }
-  if (runToken === 'totalPageCount') {
-    return context.totalPages ? String(context.totalPages) : (run.text ?? '');
-  }
-  return run.text ?? '';
 };
