+++ packages/layout-engine/painters/dom/src/runs/index.ts
@@ -1,0 +1,18 @@
+export { renderLine } from './render-line.js';
+export { renderRun } from './render-run.js';
+export { renderTextRun } from './text-run.js';
+export { renderImageRun } from './image-run.js';
+export { renderFieldAnnotationRun } from './field-annotation-run.js';
+export { renderMathRun } from './math-run.js';
+export { renderInlineTabRun, renderPositionedTabRun } from './tab-run.js';
+export { appendFormattingParagraphMark, setTextContentWithFormattingSpaceMarks } from './formatting-marks.js';
+export { sanitizeUrl, linkMetrics } from './links.js';
+export { applyRunDataAttributes } from './hash.js';
+export { resolveTrackedChangesConfig, applyTrackedChangeDecorations } from './tracked-changes.js';
+export {
+  resolveRunSdtId,
+  createInlineSdtWrapper,
+  syncInlineSdtWrapperTypography,
+  expandSdtWrapperPmRange,
+} from './sdt.js';
+export type { RenderedLineInfo, RenderLineParams, RunRenderContext, TrackedChangesRenderConfig } from './types.js';
